Splits duplicated coords (copy of geoR::dup.coords)

This function takes an object with 2-D coordinates and returns the positions of the duplicated coordinates. Also sets a method for duplicated. Helper function for plot_heatmap

dup_coords(x, ...)

Arguments

x

Two column numeric matrix or data frame

...

passed to sapply. If simplify = TRUE (default) results are returned as an array if possible (when the number of replicates are the same at each replicated location)

Value

Function and methods returns NULL if there are no duplicates locations. Otherwise, the default method returns a list where each component is a vector with the positions or the rownames, if available, of the duplicates coordinates. The method for geodata returns a data-frame with rownames equals to the positions of the duplicated coordinates, the first column is a factor indicating duplicates and the remaining are output of as.data.frame.geodata.

See also

Author

Paulo Justiniano Ribeiro Jr. paulojus@leg.ufpr.br Peter J. Diggle p.diggle@lancaster.ac.uk