Get the Grid-IDs from binary matrix

Get the grid IDs from the trimmed binary matrix, where the binary code indicates which grid cells are used in the current wind farm constellation.

getRects(trimtonOut, Grid)

Arguments

trimtonOut
Input matrix with binary values.

Grid
Grid of the considered area

Value

Returns a list of all individuals with X and Y coordinates and the grid cell ID.

Examples

## Create a random rectangular shapefile
library(sp)
Polygon1 <- Polygon(rbind(c(0, 0), c(0, 2000), c(2000, 2000), c(2000, 0)))
Polygon1 <- Polygons(list(Polygon1),1);
Polygon1 <- SpatialPolygons(list(Polygon1))
Projection <- "+proj=laea +lat_0=52 +lon_0=10 +x_0=4321000 +y_0=3210000
+ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs"
proj4string(Polygon1) <- CRS(Projection)

## Calculate a Grid and an indexed data.frame with coordinates and grid cell Ids.
Grid1 <- GridFilter(shape = Polygon1,resol = 200,prop = 1);
Grid <- Grid1[[1]]
AmountGrids <- nrow(Grid)

startsel <- StartGA(Grid,10,20);
wind <- data.frame(ws = 12, wd = 0)
wind <- list(wind, probab = 100)
fit <- fitness(selection = startsel,referenceHeight = 100, RotorHeight=100,
               SurfaceRoughness=0.3,Polygon = Polygon1, resol1 = 200,rot=20,
               dirspeed = wind, srtm_crop="",topograp=FALSE,cclRaster="")
allparks <- do.call("rbind",fit);

## SELECTION
## print the amount of Individuals selected.
## Check if the amount of Turbines is as requested.
selec6best <- selection1(fit, Grid,2, TRUE, 6, "VAR");

## CROSSOVER
## u determines the amount of crossover points,
## crossPart determines the method used (Equal/Random),
## uplimit is the maximum allowed permutations
crossOut <- crossover1(selec6best, 2, uplimit = 300, crossPart="RAN");

## MUTATION
## Variable Mutation Rate is activated if more than 2 individuals represent the
## current best solution.
mut <- mutation(a = crossOut, p = 0.3);

## TRIMTON
## After Crossover and Mutation, the amount of turbines in a windpark change
## and have to be corrected to the required amount of turbines.
mut1 <- trimton(mut = mut, nturb = 10, allparks = allparks,
               nGrids = AmountGrids, trimForce=FALSE)

## Get the new Grid-Ids and run a new fitness run.
getRectV <- getRects(mut1, Grid)
fit <- fitness(selection = getRectV,referenceHeight = 100, RotorHeight=100,
               SurfaceRoughness=0.3,Polygon = Polygon1, resol1 = 200,rot=20,
               dirspeed = wind, srtm_crop="",topograp=FALSE,cclRaster="")
head(fit)

#> $`8,17,28,31,41,50,80,81,83,93`
#>       X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#> 8  1500 1900    83.74057      6604.658           0   1     20       8
#> 17 1300 1700    83.74057      6604.658           0   1     20      17
#> 28 1500 1500    83.74057      6604.658         100   1     20      28
#> 31  100 1300    83.74057      6604.658           0   1     20      31
#> 41  100 1100    83.74057      6604.658         100   1     20      41
#> 50 1900 1100    83.74057      6604.658           0   1     20      50
#> 80 1900  500    83.74057      6604.658         100   1     20      80
#> 81  100  300    83.74057      6604.658         200   1     20      81
#> 83  500  300    83.74057      6604.658           0   1     20      83
#> 93  500  100    83.74057      6604.658         100   1     20      93
#>    Parkfitness
#> 8     6604.658
#> 17    6604.658
#> 28    6604.658
#> 31    6604.658
#> 41    6604.658
#> 50    6604.658
#> 80    6604.658
#> 81    6604.658
#> 83    6604.658
#> 93    6604.658
#> 
#> $`13,19,22,35,43,50,53,56,86,88`
#>       X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#> 13  500 1700    91.30743       7201.46           0   2     20      13
#> 19 1700 1700    91.30743       7201.46           0   2     20      19
#> 22  300 1500    91.30743       7201.46           0   2     20      22
#> 35  900 1300    91.30743       7201.46           0   2     20      35
#> 43  500 1100    91.30743       7201.46         100   2     20      43
#> 50 1900 1100    91.30743       7201.46           0   2     20      50
#> 53  500  900    91.30743       7201.46         200   2     20      53
#> 56 1100  900    91.30743       7201.46           0   2     20      56
#> 86 1100  300    91.30743       7201.46         100   2     20      86
#> 88 1500  300    91.30743       7201.46           0   2     20      88
#>    Parkfitness
#> 13     7201.46
#> 19     7201.46
#> 22     7201.46
#> 35     7201.46
#> 43     7201.46
#> 50     7201.46
#> 53     7201.46
#> 56     7201.46
#> 86     7201.46
#> 88     7201.46
#> 
#> $`19,37,42,53,63,65,76,78,84,98`
#>       X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#> 19 1700 1700    92.04609      7259.719           0   3     20      19
#> 37 1300 1300    92.04609      7259.719           0   3     20      37
#> 42  300 1100    92.04609      7259.719           0   3     20      42
#> 53  500  900    92.04609      7259.719           0   3     20      53
#> 63  500  700    92.04609      7259.719         100   3     20      63
#> 65  900  700    92.04609      7259.719           0   3     20      65
#> 76 1100  500    92.04609      7259.719           0   3     20      76
#> 78 1500  500    92.04609      7259.719           0   3     20      78
#> 84  700  300    92.04609      7259.719           0   3     20      84
#> 98 1500  100    92.04609      7259.719         100   3     20      98
#>    Parkfitness
#> 19    7259.719
#> 37    7259.719
#> 42    7259.719
#> 53    7259.719
#> 63    7259.719
#> 65    7259.719
#> 76    7259.719
#> 78    7259.719
#> 84    7259.719
#> 98    7259.719
#> 
#> $`4,5,14,21,32,52,55,56,57,100`
#>        X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#> 4    700 1900    91.20199      7193.144           0   4     20       4
#> 5    900 1900    91.20199      7193.144           0   4     20       5
#> 14   700 1700    91.20199      7193.144         100   4     20      14
#> 21   100 1500    91.20199      7193.144           0   4     20      21
#> 32   300 1300    91.20199      7193.144           0   4     20      32
#> 52   300  900    91.20199      7193.144         100   4     20      52
#> 55   900  900    91.20199      7193.144         100   4     20      55
#> 56  1100  900    91.20199      7193.144           0   4     20      56
#> 57  1300  900    91.20199      7193.144           0   4     20      57
#> 100 1900  100    91.20199      7193.144           0   4     20     100
#>     Parkfitness
#> 4      7193.144
#> 5      7193.144
#> 14     7193.144
#> 21     7193.144
#> 32     7193.144
#> 52     7193.144
#> 55     7193.144
#> 56     7193.144
#> 57     7193.144
#> 100    7193.144
#> 
#> $`5,19,20,33,44,54,83,89,91,94`
#>       X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#> 5   900 1900    92.10061      7264.019           0   5     20       5
#> 19 1700 1700    92.10061      7264.019           0   5     20      19
#> 20 1900 1700    92.10061      7264.019           0   5     20      20
#> 33  500 1300    92.10061      7264.019           0   5     20      33
#> 44  700 1100    92.10061      7264.019           0   5     20      44
#> 54  700  900    92.10061      7264.019         100   5     20      54
#> 83  500  300    92.10061      7264.019         100   5     20      83
#> 89 1700  300    92.10061      7264.019         100   5     20      89
#> 91  100  100    92.10061      7264.019           0   5     20      91
#> 94  700  100    92.10061      7264.019         200   5     20      94
#>    Parkfitness
#> 5     7264.019
#> 19    7264.019
#> 20    7264.019
#> 33    7264.019
#> 44    7264.019
#> 54    7264.019
#> 83    7264.019
#> 89    7264.019
#> 91    7264.019
#> 94    7264.019
#> 
#> $`8,15,22,29,32,47,68,79,80,100`
#>        X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#> 8   1500 1900     90.5677      7143.117           0   6     20       8
#> 15   900 1700     90.5677      7143.117           0   6     20      15
#> 22   300 1500     90.5677      7143.117           0   6     20      22
#> 29  1700 1500     90.5677      7143.117           0   6     20      29
#> 32   300 1300     90.5677      7143.117         100   6     20      32
#> 47  1300 1100     90.5677      7143.117           0   6     20      47
#> 68  1500  700     90.5677      7143.117         100   6     20      68
#> 79  1700  500     90.5677      7143.117         100   6     20      79
#> 80  1900  500     90.5677      7143.117           0   6     20      80
#> 100 1900  100     90.5677      7143.117         100   6     20     100
#>     Parkfitness
#> 8      7143.117
#> 15     7143.117
#> 22     7143.117
#> 29     7143.117
#> 32     7143.117
#> 47     7143.117
#> 68     7143.117
#> 79     7143.117
#> 80     7143.117
#> 100    7143.117
#>