Plot an interpolated viewshed

Plot an interpolated view of the viewshed analysis

interpol_view(res, plot = TRUE, breakseq, breakform = NULL,
  plotDEM = FALSE, fun = mean, pal = NULL, ...)

Arguments

res
The result list from viewshed.

plot
Should the result be plotted? Default is TRUE

breakseq
The breaks for value plotting. By default, 5 equal intervals are generated.

breakform
If 'breakseq' is missing, a sampling function to calculate the breaks, like quantile, fivenum, etc.

plotDEM
Plot the DEM? Default is FALSE

fun
Function used for rasterize. Default is mean

pal
A color palette

...
Arguments passed on to plot.

Value

An interpolated raster

See also

Examples

# NOT RUN {
library(sp)
library(raster)
Polygon1 <- Polygon(rbind(c(4488182, 2667172), c(4488182, 2669343),
                          c(4499991, 2669343), c(4499991, 2667172)))
Polygon1 <- Polygons(list(Polygon1), 1);
Polygon1 <- SpatialPolygons(list(Polygon1))
Projection <- "+proj=laea +lat_0=52 +lon_0=10 +x_0=4321000 +y_0=3210000
+ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs"
proj4string(Polygon1) <- CRS(Projection)
DEM_meter <- getDEM(Polygon1)

turbloc = spsample(DEM_meter[[2]], 10, type = "random");
res <- viewshed(r = DEM_meter[[1]], shape=DEM_meter[[2]],
                turbine_locs = turbloc,  h1=1.8, h2=50)
interpol_view(res, plotDEM = T)

interpol_view(res, breakseq = seq(0,max(colSums(res$Result)),1))
interpol_view(res, plotDEM = F, breakform = quantile)
interpol_view(res, breakform = factor)

## Different color palettes
interpol_view(res, plotDEM = T, pal=topo.colors)
interpol_view(res, plotDEM = T, pal=colorRampPalette(c("white","purple")))

## ... Arguments are past on to the raster plot method
interpol_view(res, plotDEM = T, alpha=0.5)
interpol_view(res, plotDEM = F, breakseq = seq(0,10,1), colNA="black")

# }