Deprecated use plot_windrose instead.

plotWindrose is replaced by plot_windrose.

plotWindrose(data, spd, dir, spdres = 2, dirres = 10, spdmin = 1,
  spdmax = 30, palette = "YlGnBu", spdseq = NULL, plotit = TRUE)

Arguments

data
A data.frame containing the wind information

spd
The column of the wind speeds in "data"

dir
The column of the wind directions in "data"

spdres
The increment of the wind speed legend. Default is 2

dirres
The size of the wind sectors. Default is 10

spdmin
Minimum wind speed. Default is 1

spdmax
Maximal wind speed. Default is 30

palette
A color palette used for drawing the wind rose

spdseq
A wind speed sequence, that is used for plotting

plotit
Should the windrose be plotted? Default is TRUE