Check Input Selection Method

Checks whether the input for selection is given correctly. If not, a message is prompted which asks to input one of the 2 available selection methods. The available inputs are "F" and "V". "F" refers to a fixed percentage of 50 percentage, based on the development of the population fitness values.

readintegerSel()

Value

Returns the selected selection method (character)

See also

Examples

if (FALSE) {
  readintegerSel()
}