Calculate visibility

Calculate visibility for given points in a given area.

viewshed(r, shape, turbine_locs, h1 = 0, h2 = 0, plot = FALSE, ...)

Arguments

r
A DEM raster

shape
A Simpüle Feature Polygon of the windfarm area

turbine_locs
Coordinates or SpatialPoint representing the wind turbines

h1
A numeric or numeric vector giving the extra height offsets for Point 1

h2
A numeric giving the extra height offset for Point 2

plot
Plot the process? Default is FALSE

...
Additional arguments passed to the raster plot

Value

A list of 5, containing the boolean result for every cell, the raster cell points, a SimpleFeature Polygon of the given area and the DEM raster

See also

Examples

if (FALSE) {
library(sf)
Polygon1 <- sf::st_as_sf(sf::st_sfc(
  sf::st_polygon(list(cbind(
    c(4496482, 4496482, 4499991, 4499991, 4496482),
    c(2666272, 2669343, 2669343, 2666272, 2666272)))),
  crs = 3035
))
DEM_meter <- getDEM(Polygon1)

turbloc = st_sample(DEM_meter[[2]], 10, type = "random");
res <- viewshed(r = DEM_meter[[1]], shape=DEM_meter[[2]], 
                turbine_locs = turbloc[1:3,],
                h1 = sample(c(20:40), 3, TRUE),
                h2 = 50, plot = TRUE)
}