Deprecated use calculate_energy instead.

calculateEn is replaced by calculate_energy.

calculateEn(sel, referenceHeight, RotorHeight, SurfaceRoughness, wnkl,
  distanz, polygon1, resol, RotorR, dirSpeed, srtm_crop, topograp,
  cclRaster, weibull, plotit = FALSE)

Arguments

sel
A data.frame of an individual of the current population (data.frame)

referenceHeight
The height at which the incoming wind speeds were measured (numeric)

RotorHeight
The desired height of the turbines

SurfaceRoughness
A surface roughness length of the considered area in m. If the terrain effect model is activated, a surface roughness will be calculated for every grid cell with the elevation and land cover information

wnkl
Indicates the angle at which no wake influences are considered (numeric)

distanz
Indicates the distance after which the wake effects are considered to be eliminated

polygon1
The considered area as shapefile

resol
The resolution of the grid in meter

RotorR
The desired rotor radius in meter

dirSpeed
The wind speed and direction data.frame

srtm_crop
A list of 3 raster, with 1) the elevation, 2) an orographic and 3) a terrain raster. Calculated in genetic_algorithm

topograp
Logical value that indicates whether the terrain effect model is activated (TRUE) or deactivated (FALSE)

cclRaster
A Corine Land Cover raster that has to be downloaded previously. See also the details at windfarmGA The raster will only be used when the terrain effect model is activated. (raster)

weibull
A raster representing the estimated wind speeds

plotit
Logical value. If TRUE, the process will be plotted. Default is FALSE.