Crossover Method

The crossover method of the genetic algorithm, which takes the selected individuals after the selection function and produces new offsprings through permutation.

crossover(se6, u, uplimit, crossPart, verbose, seed)

Arguments

se6
The selected individuals. The output of selection (list)

u
The crossover point rate. (numeric)

uplimit
The upper limit of allowed permutations. The current algorithm has an upper bound of 300 permutations. (numeric)

crossPart
The crossover method. Either "EQU" or "RAN". (character)

verbose
If TRUE, will print out further information.

seed
Set a seed for comparability. Default is NULL

Value

Returns a binary coded matrix of all permutations and all grid cells, 0 indicates no turbine and 1 indicates a turbine in the grid cell. (matrix)

See also

Examples

## Create two random parents with an index and random binary values
Parents <- data.frame(
  ID = 1:20,
  bin = sample(c(0,1),20, replace = TRUE, prob = c(70,30)),
  bin.1 = sample(c(0,1),20, replace=TRUE,prob = c(30,70)))

## Create random Fitness values for both individuals
FitParents <- data.frame(ID = 1, Fitness = 1000, Fitness.1 = 20)

## Assign both values to a list
CrossSampl <- list(Parents,FitParents);

## Cross their data at equal locations with 2 crossover parts
crossover(CrossSampl, u = 1.1, uplimit = 300, crossPart = "EQU")

#>       [,1] [,2] [,3] [,4]
#>  [1,]    1    1    0    0
#>  [2,]    1    1    1    1
#>  [3,]    0    0    0    0
#>  [4,]    1    1    0    0
#>  [5,]    0    0    0    0
#>  [6,]    0    0    0    0
#>  [7,]    1    1    0    0
#>  [8,]    1    1    0    0
#>  [9,]    1    1    1    1
#> [10,]    1    1    1    1
#> [11,]    1    1    1    1
#> [12,]    1    0    1    0
#> [13,]    1    0    1    0
#> [14,]    1    0    1    0
#> [15,]    1    0    1    0
#> [16,]    1    0    1    0
#> [17,]    0    0    0    0
#> [18,]    0    0    0    0
#> [19,]    1    0    1    0
#> [20,]    1    1    1    1

## with 3 crossover parts and equal locations
crossover(CrossSampl, u = 2.5, uplimit = 300, crossPart = "EQU")

#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8]
#>  [1,]    1    1    1    1    0    0    0    0
#>  [2,]    1    1    1    1    1    1    1    1
#>  [3,]    0    0    0    0    0    0    0    0
#>  [4,]    1    1    1    1    0    0    0    0
#>  [5,]    0    0    0    0    0    0    0    0
#>  [6,]    0    0    0    0    0    0    0    0
#>  [7,]    1    1    1    1    0    0    0    0
#>  [8,]    1    1    0    0    1    1    0    0
#>  [9,]    1    1    1    1    1    1    1    1
#> [10,]    1    1    1    1    1    1    1    1
#> [11,]    1    1    1    1    1    1    1    1
#> [12,]    1    1    0    0    1    1    0    0
#> [13,]    1    1    0    0    1    1    0    0
#> [14,]    1    1    0    0    1    1    0    0
#> [15,]    1    0    1    0    1    0    1    0
#> [16,]    1    0    1    0    1    0    1    0
#> [17,]    0    0    0    0    0    0    0    0
#> [18,]    0    0    0    0    0    0    0    0
#> [19,]    1    0    1    0    1    0    1    0
#> [20,]    1    1    1    1    1    1    1    1

## or with random locations and 5 crossover parts
crossover(CrossSampl, u = 4.9, uplimit = 300, crossPart = "RAN")

#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#>  [2,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#>  [3,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#>  [4,]    0    0    0    0    0    0    0    0    1     1     1     1     1
#>  [5,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#>  [6,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#>  [7,]    0    0    0    0    1    1    1    1    0     0     0     0     1
#>  [8,]    0    0    1    1    0    0    1    1    0     0     1     1     0
#>  [9,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#> [10,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#> [11,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#> [12,]    0    0    1    1    0    0    1    1    0     0     1     1     0
#> [13,]    0    0    1    1    0    0    1    1    0     0     1     1     0
#> [14,]    0    0    1    1    0    0    1    1    0     0     1     1     0
#> [15,]    0    1    0    1    0    1    0    1    0     1     0     1     0
#> [16,]    0    1    0    1    0    1    0    1    0     1     0     1     0
#> [17,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [18,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [19,]    0    1    0    1    0    1    0    1    0     1     0     1     0
#> [20,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22] [,23] [,24] [,25]
#>  [1,]     0     0     0     1     1     1     1     1     1     1     1     1
#>  [2,]     1     1     1     1     1     1     1     1     1     1     1     1
#>  [3,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [4,]     1     1     1     0     0     0     0     0     0     0     0     1
#>  [5,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [6,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [7,]     1     1     1     0     0     0     0     1     1     1     1     0
#>  [8,]     0     1     1     0     0     1     1     0     0     1     1     0
#>  [9,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [10,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [11,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [12,]     0     1     1     0     0     1     1     0     0     1     1     0
#> [13,]     0     1     1     0     0     1     1     0     0     1     1     0
#> [14,]     0     1     1     0     0     1     1     0     0     1     1     0
#> [15,]     1     0     1     0     1     0     1     0     1     0     1     0
#> [16,]     1     0     1     0     1     0     1     0     1     0     1     0
#> [17,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [18,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [19,]     1     0     1     0     1     0     1     0     1     0     1     0
#> [20,]     1     1     1     1     1     1     1     1     1     1     1     1
#>       [,26] [,27] [,28] [,29] [,30] [,31] [,32]
#>  [1,]     1     1     1     1     1     1     1
#>  [2,]     1     1     1     1     1     1     1
#>  [3,]     0     0     0     0     0     0     0
#>  [4,]     1     1     1     1     1     1     1
#>  [5,]     0     0     0     0     0     0     0
#>  [6,]     0     0     0     0     0     0     0
#>  [7,]     0     0     0     1     1     1     1
#>  [8,]     0     1     1     0     0     1     1
#>  [9,]     1     1     1     1     1     1     1
#> [10,]     1     1     1     1     1     1     1
#> [11,]     1     1     1     1     1     1     1
#> [12,]     0     1     1     0     0     1     1
#> [13,]     0     1     1     0     0     1     1
#> [14,]     0     1     1     0     0     1     1
#> [15,]     1     0     1     0     1     0     1
#> [16,]     1     0     1     0     1     0     1
#> [17,]     0     0     0     0     0     0     0
#> [18,]     0     0     0     0     0     0     0
#> [19,]     1     0     1     0     1     0     1
#> [20,]     1     1     1     1     1     1     1