Get the Grid-IDs from binary matrix

Retrieve the grid ID's from the binary matrix, where the binary code indicates which grid cells are used in the current wind farm constellation.

get_grids(trimtonOut, Grid)

Arguments

trimtonOut

Input matrix with binary values

Grid

Grid of the considered area

Value

Returns a list of all individuals with X and Y coordinates and the grid cell ID.

See also

Examples

# \donttest{
## Create a random rectangular shapefile
library(sf)
Polygon1 <- sf::st_as_sf(sf::st_sfc(
 sf::st_polygon(list(cbind(
   c(0, 0, 2000, 2000, 0),
   c(0, 2000, 2000, 0, 0)))),
crs = 3035
))

## Calculate a Grid and an indexed data.frame with coordinates and grid cell Ids.
Grid1 <- grid_area(shape = Polygon1, size = 200, prop = 1)
Grid <- Grid1[[1]]
AmountGrids <- nrow(Grid)

startsel <- init_population(Grid,10,20)
wind <- data.frame(ws = 12, wd = 0)
wind <- list(wind, probab = 100)
fit <- fitness(selection = startsel,referenceHeight = 100, RotorHeight=100,
               SurfaceRoughness=0.3,Polygon = Polygon1, resol1 = 200,rot=20,
               dirspeed = wind, srtm_crop="",topograp=FALSE,cclRaster="")
allparks <- do.call("rbind",fit)

## SELECTION
## print the amount of Individuals selected.
## Check if the amount of Turbines is as requested.
selec6best <- selection(fit, Grid,2, TRUE, 6, "VAR")

## CROSSOVER
## u determines the amount of crossover points,
## crossPart determines the method used (Equal/Random),
## uplimit is the maximum allowed permutations
crossOut <- crossover(selec6best, 2, uplimit = 300, crossPart="RAN")

## MUTATION
## Variable Mutation Rate is activated if more than 2 individuals represent the
## current best solution.
mut <- mutation(a = crossOut, p = 0.3)

## TRIMTON
## After Crossover and Mutation, the amount of turbines in a windpark change
## and have to be corrected to the required amount of turbines.
mut1 <- trimton(mut = mut, nturb = 10, allparks = allparks,
               nGrids = AmountGrids, trimForce=FALSE)

## Get the new Grid-Ids and run a new fitness run.
getRectV <- get_grids(mut1, Grid)
fit <- fitness(selection = getRectV,referenceHeight = 100, RotorHeight=100,
               SurfaceRoughness=0.3,Polygon = Polygon1, resol1 = 200,rot=20,
               dirspeed = wind, srtm_crop="",topograp=FALSE,cclRaster="")
head(fit)
#> $`8,12,15,54,62,63,65,74,78,99`
#>          X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#>  [1,] 1500  100    94.99759      7492.504         100   1     20       8
#>  [2,]  300  300    94.99759      7492.504         100   1     20      12
#>  [3,]  900  300    94.99759      7492.504         100   1     20      15
#>  [4,]  700 1100    94.99759      7492.504         100   1     20      54
#>  [5,]  300 1300    94.99759      7492.504           0   1     20      62
#>  [6,]  500 1300    94.99759      7492.504           0   1     20      63
#>  [7,]  900 1300    94.99759      7492.504           0   1     20      65
#>  [8,]  700 1500    94.99759      7492.504           0   1     20      74
#>  [9,] 1500 1500    94.99759      7492.504           0   1     20      78
#> [10,] 1700 1900    94.99759      7492.504           0   1     20      99
#>       Parkfitness
#>  [1,]    7492.504
#>  [2,]    7492.504
#>  [3,]    7492.504
#>  [4,]    7492.504
#>  [5,]    7492.504
#>  [6,]    7492.504
#>  [7,]    7492.504
#>  [8,]    7492.504
#>  [9,]    7492.504
#> [10,]    7492.504
#> 
#> $`12,14,21,43,61,68,78,89,90,98`
#>          X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#>  [1,]  300  300    90.70866      7154.235           0   2     20      12
#>  [2,]  700  300    90.70866      7154.235           0   2     20      14
#>  [3,]  100  500    90.70866      7154.235         100   2     20      21
#>  [4,]  500  900    90.70866      7154.235           0   2     20      43
#>  [5,]  100 1300    90.70866      7154.235           0   2     20      61
#>  [6,] 1500 1300    90.70866      7154.235         200   2     20      68
#>  [7,] 1500 1500    90.70866      7154.235         100   2     20      78
#>  [8,] 1700 1700    90.70866      7154.235           0   2     20      89
#>  [9,] 1900 1700    90.70866      7154.235           0   2     20      90
#> [10,] 1500 1900    90.70866      7154.235           0   2     20      98
#>       Parkfitness
#>  [1,]    7154.235
#>  [2,]    7154.235
#>  [3,]    7154.235
#>  [4,]    7154.235
#>  [5,]    7154.235
#>  [6,]    7154.235
#>  [7,]    7154.235
#>  [8,]    7154.235
#>  [9,]    7154.235
#> [10,]    7154.235
#> 
#> $`9,14,23,66,68,77,83,88,96,99`
#>          X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#>  [1,] 1700  100    94.47836      7451.552         100   3     20       9
#>  [2,]  700  300    94.47836      7451.552           0   3     20      14
#>  [3,]  500  500    94.47836      7451.552         100   3     20      23
#>  [4,] 1100 1300    94.47836      7451.552         100   3     20      66
#>  [5,] 1500 1300    94.47836      7451.552         100   3     20      68
#>  [6,] 1300 1500    94.47836      7451.552           0   3     20      77
#>  [7,]  500 1700    94.47836      7451.552           0   3     20      83
#>  [8,] 1500 1700    94.47836      7451.552           0   3     20      88
#>  [9,] 1100 1900    94.47836      7451.552           0   3     20      96
#> [10,] 1700 1900    94.47836      7451.552           0   3     20      99
#>       Parkfitness
#>  [1,]    7451.552
#>  [2,]    7451.552
#>  [3,]    7451.552
#>  [4,]    7451.552
#>  [5,]    7451.552
#>  [6,]    7451.552
#>  [7,]    7451.552
#>  [8,]    7451.552
#>  [9,]    7451.552
#> [10,]    7451.552
#> 
#> $`1,17,55,59,61,67,78,82,84,89`
#>          X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#>  [1,]  100  100    96.77776      7632.907         100   4     20       1
#>  [2,] 1300  300    96.77776      7632.907         100   4     20      17
#>  [3,]  900 1100    96.77776      7632.907           0   4     20      55
#>  [4,] 1700 1100    96.77776      7632.907         100   4     20      59
#>  [5,]  100 1300    96.77776      7632.907           0   4     20      61
#>  [6,] 1300 1300    96.77776      7632.907           0   4     20      67
#>  [7,] 1500 1500    96.77776      7632.907           0   4     20      78
#>  [8,]  300 1700    96.77776      7632.907           0   4     20      82
#>  [9,]  700 1700    96.77776      7632.907           0   4     20      84
#> [10,] 1700 1700    96.77776      7632.907           0   4     20      89
#>       Parkfitness
#>  [1,]    7632.907
#>  [2,]    7632.907
#>  [3,]    7632.907
#>  [4,]    7632.907
#>  [5,]    7632.907
#>  [6,]    7632.907
#>  [7,]    7632.907
#>  [8,]    7632.907
#>  [9,]    7632.907
#> [10,]    7632.907
#> 
#> $`11,12,22,31,47,49,53,62,96,97`
#>          X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#>  [1,]  100  300    89.98965      7097.526         100   5     20      11
#>  [2,]  300  300    89.98965      7097.526         200   5     20      12
#>  [3,]  300  500    89.98965      7097.526         100   5     20      22
#>  [4,]  100  700    89.98965      7097.526           0   5     20      31
#>  [5,] 1300  900    89.98965      7097.526         100   5     20      47
#>  [6,] 1700  900    89.98965      7097.526           0   5     20      49
#>  [7,]  500 1100    89.98965      7097.526           0   5     20      53
#>  [8,]  300 1300    89.98965      7097.526           0   5     20      62
#>  [9,] 1100 1900    89.98965      7097.526           0   5     20      96
#> [10,] 1300 1900    89.98965      7097.526           0   5     20      97
#>       Parkfitness
#>  [1,]    7097.526
#>  [2,]    7097.526
#>  [3,]    7097.526
#>  [4,]    7097.526
#>  [5,]    7097.526
#>  [6,]    7097.526
#>  [7,]    7097.526
#>  [8,]    7097.526
#>  [9,]    7097.526
#> [10,]    7097.526
#> 
#> $`2,3,20,21,23,28,44,63,77,84`
#>          X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#>  [1,]  300  100     94.7742      7474.886           0   6     20       2
#>  [2,]  500  100     94.7742      7474.886         200   6     20       3
#>  [3,] 1900  300     94.7742      7474.886           0   6     20      20
#>  [4,]  100  500     94.7742      7474.886           0   6     20      21
#>  [5,]  500  500     94.7742      7474.886         100   6     20      23
#>  [6,] 1500  500     94.7742      7474.886           0   6     20      28
#>  [7,]  700  900     94.7742      7474.886         100   6     20      44
#>  [8,]  500 1300     94.7742      7474.886           0   6     20      63
#>  [9,] 1300 1500     94.7742      7474.886           0   6     20      77
#> [10,]  700 1700     94.7742      7474.886           0   6     20      84
#>       Parkfitness
#>  [1,]    7474.886
#>  [2,]    7474.886
#>  [3,]    7474.886
#>  [4,]    7474.886
#>  [5,]    7474.886
#>  [6,]    7474.886
#>  [7,]    7474.886
#>  [8,]    7474.886
#>  [9,]    7474.886
#> [10,]    7474.886
#> 

# }