Get the Grid-IDs from binary matrix

Get the grid IDs from the trimmed binary matrix, where the binary code indicates which grid cells are used in the current wind farm constellation.

get_grids(trimtonOut, Grid)

Arguments

trimtonOut
Input matrix with binary values.

Grid
Grid of the considered area

Value

Returns a list of all individuals with X and Y coordinates and the grid cell ID.

See also

Examples

# \donttest{
## Create a random rectangular shapefile
library(sf)
Polygon1 <- sf::st_as_sf(sf::st_sfc(
 sf::st_polygon(list(cbind(
   c(0, 0, 2000, 2000, 0),
   c(0, 2000, 2000, 0, 0)))),
crs = 3035
))

## Calculate a Grid and an indexed data.frame with coordinates and grid cell Ids.
Grid1 <- grid_area(shape = Polygon1, size = 200, prop = 1)
Grid <- Grid1[[1]]
AmountGrids <- nrow(Grid)

startsel <- init_population(Grid,10,20)
wind <- data.frame(ws = 12, wd = 0)
wind <- list(wind, probab = 100)
fit <- fitness(selection = startsel,referenceHeight = 100, RotorHeight=100,
               SurfaceRoughness=0.3,Polygon = Polygon1, resol1 = 200,rot=20,
               dirspeed = wind, srtm_crop="",topograp=FALSE,cclRaster="")

#> Error in subset.matrix(res, subset = res[, "alpha"] < wnkl & res[, "Laenge_B"] <     distanz): object 'distanz' not found

allparks <- do.call("rbind",fit)

#> Error in do.call("rbind", fit): object 'fit' not found

## SELECTION
## print the amount of Individuals selected.
## Check if the amount of Turbines is as requested.
selec6best <- selection(fit, Grid,2, TRUE, 6, "VAR")

#> Error in do.call("rbind", fit): object 'fit' not found

## CROSSOVER
## u determines the amount of crossover points,
## crossPart determines the method used (Equal/Random),
## uplimit is the maximum allowed permutations
crossOut <- crossover(selec6best, 2, uplimit = 300, crossPart="RAN")

#> Error in crossover(selec6best, 2, uplimit = 300, crossPart = "RAN"): object 'selec6best' not found

## MUTATION
## Variable Mutation Rate is activated if more than 2 individuals represent the
## current best solution.
mut <- mutation(a = crossOut, p = 0.3)

#> Error in runif(n = length(a), min = 0, max = 1): object 'crossOut' not found

## TRIMTON
## After Crossover and Mutation, the amount of turbines in a windpark change
## and have to be corrected to the required amount of turbines.
mut1 <- trimton(mut = mut, nturb = 10, allparks = allparks,
               nGrids = AmountGrids, trimForce=FALSE)

#> Error in ncol(x): object 'allparks' not found

## Get the new Grid-Ids and run a new fitness run.
getRectV <- get_grids(mut1, Grid)

#> Error in get_grids(mut1, Grid): object 'mut1' not found

fit <- fitness(selection = getRectV,referenceHeight = 100, RotorHeight=100,
               SurfaceRoughness=0.3,Polygon = Polygon1, resol1 = 200,rot=20,
               dirspeed = wind, srtm_crop="",topograp=FALSE,cclRaster="")

#> Error in fitness(selection = getRectV, referenceHeight = 100, RotorHeight = 100,     SurfaceRoughness = 0.3, Polygon = Polygon1, resol1 = 200,     rot = 20, dirspeed = wind, srtm_crop = "", topograp = FALSE,     cclRaster = ""): object 'getRectV' not found

head(fit)

#> Error in head(fit): object 'fit' not found

# }