Get the Grid-IDs from binary matrix

Retrieve the grid ID's from the binary matrix, where the binary code indicates which grid cells are used in the current wind farm constellation.

get_grids(trimtonOut, Grid)

Arguments

trimtonOut

Input matrix with binary values

Grid

Grid of the considered area

Value

Returns a list of all individuals with X and Y coordinates and the grid cell ID.

See also

Examples

# \donttest{
## Create a random rectangular shapefile
library(sf)
Polygon1 <- sf::st_as_sf(sf::st_sfc(
 sf::st_polygon(list(cbind(
   c(0, 0, 2000, 2000, 0),
   c(0, 2000, 2000, 0, 0)))),
crs = 3035
))

## Calculate a Grid and an indexed data.frame with coordinates and grid cell Ids.
Grid1 <- grid_area(shape = Polygon1, size = 200, prop = 1)
Grid <- Grid1[[1]]
AmountGrids <- nrow(Grid)

startsel <- init_population(Grid,10,20)
wind <- data.frame(ws = 12, wd = 0)
wind <- list(wind, probab = 100)
fit <- fitness(selection = startsel,referenceHeight = 100, RotorHeight=100,
               SurfaceRoughness=0.3,Polygon = Polygon1, resol1 = 200,rot=20,
               dirspeed = wind, srtm_crop="",topograp=FALSE,cclRaster="")
allparks <- do.call("rbind",fit)

## SELECTION
## print the amount of Individuals selected.
## Check if the amount of Turbines is as requested.
selec6best <- selection(fit, Grid,2, TRUE, 6, "VAR")

## CROSSOVER
## u determines the amount of crossover points,
## crossPart determines the method used (Equal/Random),
## uplimit is the maximum allowed permutations
crossOut <- crossover(selec6best, 2, uplimit = 300, crossPart="RAN")

## MUTATION
## Variable Mutation Rate is activated if more than 2 individuals represent the
## current best solution.
mut <- mutation(a = crossOut, p = 0.3)

## TRIMTON
## After Crossover and Mutation, the amount of turbines in a windpark change
## and have to be corrected to the required amount of turbines.
mut1 <- trimton(mut = mut, nturb = 10, allparks = allparks,
               nGrids = AmountGrids, trimForce=FALSE)

## Get the new Grid-Ids and run a new fitness run.
getRectV <- get_grids(mut1, Grid)
fit <- fitness(selection = getRectV,referenceHeight = 100, RotorHeight=100,
               SurfaceRoughness=0.3,Polygon = Polygon1, resol1 = 200,rot=20,
               dirspeed = wind, srtm_crop="",topograp=FALSE,cclRaster="")
head(fit)
#> $`4,20,36,64,65,67,70,80,96,98`
#>          X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#>  [1,]  700  100    92.55268      7299.673         100   1     20       4
#>  [2,] 1900  300    92.55268      7299.673         200   1     20      20
#>  [3,] 1100  700    92.55268      7299.673         100   1     20      36
#>  [4,]  700 1300    92.55268      7299.673           0   1     20      64
#>  [5,]  900 1300    92.55268      7299.673           0   1     20      65
#>  [6,] 1300 1300    92.55268      7299.673           0   1     20      67
#>  [7,] 1900 1300    92.55268      7299.673         100   1     20      70
#>  [8,] 1900 1500    92.55268      7299.673           0   1     20      80
#>  [9,] 1100 1900    92.55268      7299.673           0   1     20      96
#> [10,] 1500 1900    92.55268      7299.673           0   1     20      98
#>       Parkfitness
#>  [1,]    7299.673
#>  [2,]    7299.673
#>  [3,]    7299.673
#>  [4,]    7299.673
#>  [5,]    7299.673
#>  [6,]    7299.673
#>  [7,]    7299.673
#>  [8,]    7299.673
#>  [9,]    7299.673
#> [10,]    7299.673
#> 
#> $`10,20,26,42,56,62,64,66,77,98`
#>          X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#>  [1,] 1900  100    84.84926      6692.101         100   2     20      10
#>  [2,] 1900  300    84.84926      6692.101           0   2     20      20
#>  [3,] 1100  500    84.84926      6692.101         200   2     20      26
#>  [4,]  300  900    84.84926      6692.101         100   2     20      42
#>  [5,] 1100 1100    84.84926      6692.101         100   2     20      56
#>  [6,]  300 1300    84.84926      6692.101           0   2     20      62
#>  [7,]  700 1300    84.84926      6692.101           0   2     20      64
#>  [8,] 1100 1300    84.84926      6692.101           0   2     20      66
#>  [9,] 1300 1500    84.84926      6692.101           0   2     20      77
#> [10,] 1500 1900    84.84926      6692.101           0   2     20      98
#>       Parkfitness
#>  [1,]    6692.101
#>  [2,]    6692.101
#>  [3,]    6692.101
#>  [4,]    6692.101
#>  [5,]    6692.101
#>  [6,]    6692.101
#>  [7,]    6692.101
#>  [8,]    6692.101
#>  [9,]    6692.101
#> [10,]    6692.101
#> 
#> $`9,18,41,42,48,51,55,60,62,90`
#>          X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#>  [1,] 1700  100    88.55839      6984.642           0   3     20       9
#>  [2,] 1500  300    88.55839      6984.642         100   3     20      18
#>  [3,]  100  900    88.55839      6984.642         100   3     20      41
#>  [4,]  300  900    88.55839      6984.642         100   3     20      42
#>  [5,] 1500  900    88.55839      6984.642           0   3     20      48
#>  [6,]  100 1100    88.55839      6984.642           0   3     20      51
#>  [7,]  900 1100    88.55839      6984.642           0   3     20      55
#>  [8,] 1900 1100    88.55839      6984.642         100   3     20      60
#>  [9,]  300 1300    88.55839      6984.642           0   3     20      62
#> [10,] 1900 1700    88.55839      6984.642           0   3     20      90
#>       Parkfitness
#>  [1,]    6984.642
#>  [2,]    6984.642
#>  [3,]    6984.642
#>  [4,]    6984.642
#>  [5,]    6984.642
#>  [6,]    6984.642
#>  [7,]    6984.642
#>  [8,]    6984.642
#>  [9,]    6984.642
#> [10,]    6984.642
#> 
#> $`23,32,41,47,50,61,70,78,94,99`
#>          X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#>  [1,]  500  500    94.35881      7442.124           0   4     20      23
#>  [2,]  300  700    94.35881      7442.124           0   4     20      32
#>  [3,]  100  900    94.35881      7442.124         100   4     20      41
#>  [4,] 1300  900    94.35881      7442.124           0   4     20      47
#>  [5,] 1900  900    94.35881      7442.124         100   4     20      50
#>  [6,]  100 1300    94.35881      7442.124           0   4     20      61
#>  [7,] 1900 1300    94.35881      7442.124           0   4     20      70
#>  [8,] 1500 1500    94.35881      7442.124           0   4     20      78
#>  [9,]  700 1900    94.35881      7442.124           0   4     20      94
#> [10,] 1700 1900    94.35881      7442.124           0   4     20      99
#>       Parkfitness
#>  [1,]    7442.124
#>  [2,]    7442.124
#>  [3,]    7442.124
#>  [4,]    7442.124
#>  [5,]    7442.124
#>  [6,]    7442.124
#>  [7,]    7442.124
#>  [8,]    7442.124
#>  [9,]    7442.124
#> [10,]    7442.124
#> 
#> $`17,20,32,41,48,70,72,74,82,97`
#>          X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#>  [1,] 1300  300    92.19937      7271.808         100   5     20      17
#>  [2,] 1900  300    92.19937      7271.808         100   5     20      20
#>  [3,]  300  700    92.19937      7271.808         200   5     20      32
#>  [4,]  100  900    92.19937      7271.808           0   5     20      41
#>  [5,] 1500  900    92.19937      7271.808           0   5     20      48
#>  [6,] 1900 1300    92.19937      7271.808           0   5     20      70
#>  [7,]  300 1500    92.19937      7271.808         100   5     20      72
#>  [8,]  700 1500    92.19937      7271.808           0   5     20      74
#>  [9,]  300 1700    92.19937      7271.808           0   5     20      82
#> [10,] 1300 1900    92.19937      7271.808           0   5     20      97
#>       Parkfitness
#>  [1,]    7271.808
#>  [2,]    7271.808
#>  [3,]    7271.808
#>  [4,]    7271.808
#>  [5,]    7271.808
#>  [6,]    7271.808
#>  [7,]    7271.808
#>  [8,]    7271.808
#>  [9,]    7271.808
#> [10,]    7271.808
#> 
#> $`17,26,28,32,41,64,75,79,93,94`
#>          X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#>  [1,] 1300  300    98.25615      7749.509           0   6     20      17
#>  [2,] 1100  500    98.25615      7749.509           0   6     20      26
#>  [3,] 1500  500    98.25615      7749.509           0   6     20      28
#>  [4,]  300  700    98.25615      7749.509           0   6     20      32
#>  [5,]  100  900    98.25615      7749.509           0   6     20      41
#>  [6,]  700 1300    98.25615      7749.509         100   6     20      64
#>  [7,]  900 1500    98.25615      7749.509           0   6     20      75
#>  [8,] 1700 1500    98.25615      7749.509           0   6     20      79
#>  [9,]  500 1900    98.25615      7749.509           0   6     20      93
#> [10,]  700 1900    98.25615      7749.509           0   6     20      94
#>       Parkfitness
#>  [1,]    7749.509
#>  [2,]    7749.509
#>  [3,]    7749.509
#>  [4,]    7749.509
#>  [5,]    7749.509
#>  [6,]    7749.509
#>  [7,]    7749.509
#>  [8,]    7749.509
#>  [9,]    7749.509
#> [10,]    7749.509
#> 

# }