Get the Grid-IDs from binary matrix

Retrieve the grid ID's from the binary matrix, where the binary code indicates which grid cells are used in the current wind farm constellation.

Usage

get_grids(trimtonOut, Grid)

Arguments

trimtonOut

Input matrix with binary values

Grid

Grid of the considered area

Value

Returns a list of all individuals with X and Y coordinates and the grid cell ID.

See also

Examples

# \donttest{
## Create a random rectangular shapefile
library(sf)
Polygon1 <- sf::st_as_sf(sf::st_sfc(
 sf::st_polygon(list(cbind(
   c(0, 0, 2000, 2000, 0),
   c(0, 2000, 2000, 0, 0)))),
crs = 3035
))

## Calculate a Grid and an indexed data.frame with coordinates and grid cell Ids.
Grid1 <- grid_area(shape = Polygon1, size = 200, prop = 1)
Grid <- Grid1[[1]]
AmountGrids <- nrow(Grid)

startsel <- init_population(Grid,10,20)
wind <- data.frame(ws = 12, wd = 0)
wind <- list(wind, probab = 100)
fit <- fitness(selection = startsel,referenceHeight = 100, RotorHeight=100,
               SurfaceRoughness=0.3,Polygon = Polygon1, resol1 = 200,rot=20,
               dirspeed = wind, srtm_crop="",topograp=FALSE,cclRaster="")
allparks <- do.call("rbind",fit)

## SELECTION
## print the amount of Individuals selected.
## Check if the amount of Turbines is as requested.
selec6best <- selection(fit, Grid,2, TRUE, 6, "VAR")

## CROSSOVER
## u determines the amount of crossover points,
## crossPart determines the method used (Equal/Random),
## uplimit is the maximum allowed permutations
crossOut <- crossover(selec6best, 2, uplimit = 300, crossPart="RAN")

## MUTATION
## Variable Mutation Rate is activated if more than 2 individuals represent the
## current best solution.
mut <- mutation(a = crossOut, p = 0.3)

## TRIMTON
## After Crossover and Mutation, the amount of turbines in a windpark change
## and have to be corrected to the required amount of turbines.
mut1 <- trimton(mut = mut, nturb = 10, allparks = allparks,
               nGrids = AmountGrids, trimForce=FALSE)

## Get the new Grid-Ids and run a new fitness run.
getRectV <- get_grids(mut1, Grid)
fit <- fitness(selection = getRectV,referenceHeight = 100, RotorHeight=100,
               SurfaceRoughness=0.3,Polygon = Polygon1, resol1 = 200,rot=20,
               dirspeed = wind, srtm_crop="",topograp=FALSE,cclRaster="")
head(fit)
#> $`2,4,29,42,46,56,61,86,89,91`
#>          X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#>  [1,]  300  100    89.49704      7058.674         100   1     20       2
#>  [2,]  700  100    89.49704      7058.674           0   1     20       4
#>  [3,] 1700  500    89.49704      7058.674         100   1     20      29
#>  [4,]  300  900    89.49704      7058.674           0   1     20      42
#>  [5,] 1100  900    89.49704      7058.674         200   1     20      46
#>  [6,] 1100 1100    89.49704      7058.674         100   1     20      56
#>  [7,]  100 1300    89.49704      7058.674         100   1     20      61
#>  [8,] 1100 1700    89.49704      7058.674           0   1     20      86
#>  [9,] 1700 1700    89.49704      7058.674           0   1     20      89
#> [10,]  100 1900    89.49704      7058.674           0   1     20      91
#>       Parkfitness
#>  [1,]    7058.674
#>  [2,]    7058.674
#>  [3,]    7058.674
#>  [4,]    7058.674
#>  [5,]    7058.674
#>  [6,]    7058.674
#>  [7,]    7058.674
#>  [8,]    7058.674
#>  [9,]    7058.674
#> [10,]    7058.674
#> 
#> $`12,23,26,36,41,57,77,82,88,93`
#>          X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#>  [1,]  300  300    91.05888      7181.857         100   2     20      12
#>  [2,]  500  500    91.05888      7181.857         100   2     20      23
#>  [3,] 1100  500    91.05888      7181.857         100   2     20      26
#>  [4,] 1100  700    91.05888      7181.857           0   2     20      36
#>  [5,]  100  900    91.05888      7181.857           0   2     20      41
#>  [6,] 1300 1100    91.05888      7181.857         100   2     20      57
#>  [7,] 1300 1500    91.05888      7181.857           0   2     20      77
#>  [8,]  300 1700    91.05888      7181.857           0   2     20      82
#>  [9,] 1500 1700    91.05888      7181.857           0   2     20      88
#> [10,]  500 1900    91.05888      7181.857           0   2     20      93
#>       Parkfitness
#>  [1,]    7181.857
#>  [2,]    7181.857
#>  [3,]    7181.857
#>  [4,]    7181.857
#>  [5,]    7181.857
#>  [6,]    7181.857
#>  [7,]    7181.857
#>  [8,]    7181.857
#>  [9,]    7181.857
#> [10,]    7181.857
#> 
#> $`10,29,39,49,82,83,87,91,92,99`
#>          X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#>  [1,] 1900  100    83.28922       6569.06           0   3     20      10
#>  [2,] 1700  500    83.28922       6569.06         300   3     20      29
#>  [3,] 1700  700    83.28922       6569.06         200   3     20      39
#>  [4,] 1700  900    83.28922       6569.06         100   3     20      49
#>  [5,]  300 1700    83.28922       6569.06         100   3     20      82
#>  [6,]  500 1700    83.28922       6569.06           0   3     20      83
#>  [7,] 1300 1700    83.28922       6569.06           0   3     20      87
#>  [8,]  100 1900    83.28922       6569.06           0   3     20      91
#>  [9,]  300 1900    83.28922       6569.06           0   3     20      92
#> [10,] 1700 1900    83.28922       6569.06           0   3     20      99
#>       Parkfitness
#>  [1,]     6569.06
#>  [2,]     6569.06
#>  [3,]     6569.06
#>  [4,]     6569.06
#>  [5,]     6569.06
#>  [6,]     6569.06
#>  [7,]     6569.06
#>  [8,]     6569.06
#>  [9,]     6569.06
#> [10,]     6569.06
#> 
#> $`8,29,56,60,70,75,81,97,99,100`
#>          X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#>  [1,] 1500  100    92.55767      7300.067           0   4     20       8
#>  [2,] 1700  500    92.55767      7300.067         100   4     20      29
#>  [3,] 1100 1100    92.55767      7300.067           0   4     20      56
#>  [4,] 1900 1100    92.55767      7300.067         200   4     20      60
#>  [5,] 1900 1300    92.55767      7300.067         100   4     20      70
#>  [6,]  900 1500    92.55767      7300.067           0   4     20      75
#>  [7,]  100 1700    92.55767      7300.067           0   4     20      81
#>  [8,] 1300 1900    92.55767      7300.067           0   4     20      97
#>  [9,] 1700 1900    92.55767      7300.067           0   4     20      99
#> [10,] 1900 1900    92.55767      7300.067           0   4     20     100
#>       Parkfitness
#>  [1,]    7300.067
#>  [2,]    7300.067
#>  [3,]    7300.067
#>  [4,]    7300.067
#>  [5,]    7300.067
#>  [6,]    7300.067
#>  [7,]    7300.067
#>  [8,]    7300.067
#>  [9,]    7300.067
#> [10,]    7300.067
#> 
#> $`28,43,48,49,51,53,69,88,91,93`
#>          X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#>  [1,] 1500  500    85.60796       6751.94         200   5     20      28
#>  [2,]  500  900    85.60796       6751.94         200   5     20      43
#>  [3,] 1500  900    85.60796       6751.94         100   5     20      48
#>  [4,] 1700  900    85.60796       6751.94         100   5     20      49
#>  [5,]  100 1100    85.60796       6751.94         100   5     20      51
#>  [6,]  500 1100    85.60796       6751.94         100   5     20      53
#>  [7,] 1700 1300    85.60796       6751.94           0   5     20      69
#>  [8,] 1500 1700    85.60796       6751.94           0   5     20      88
#>  [9,]  100 1900    85.60796       6751.94           0   5     20      91
#> [10,]  500 1900    85.60796       6751.94           0   5     20      93
#>       Parkfitness
#>  [1,]     6751.94
#>  [2,]     6751.94
#>  [3,]     6751.94
#>  [4,]     6751.94
#>  [5,]     6751.94
#>  [6,]     6751.94
#>  [7,]     6751.94
#>  [8,]     6751.94
#>  [9,]     6751.94
#> [10,]     6751.94
#> 
#> $`30,31,50,68,71,74,77,82,91,95`
#>          X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#>  [1,] 1900  500    93.03555      7337.757         100   6     20      30
#>  [2,]  100  700    93.03555      7337.757         200   6     20      31
#>  [3,] 1900  900    93.03555      7337.757           0   6     20      50
#>  [4,] 1500 1300    93.03555      7337.757           0   6     20      68
#>  [5,]  100 1500    93.03555      7337.757         100   6     20      71
#>  [6,]  700 1500    93.03555      7337.757           0   6     20      74
#>  [7,] 1300 1500    93.03555      7337.757           0   6     20      77
#>  [8,]  300 1700    93.03555      7337.757           0   6     20      82
#>  [9,]  100 1900    93.03555      7337.757           0   6     20      91
#> [10,]  900 1900    93.03555      7337.757           0   6     20      95
#>       Parkfitness
#>  [1,]    7337.757
#>  [2,]    7337.757
#>  [3,]    7337.757
#>  [4,]    7337.757
#>  [5,]    7337.757
#>  [6,]    7337.757
#>  [7,]    7337.757
#>  [8,]    7337.757
#>  [9,]    7337.757
#> [10,]    7337.757
#> 

# }