Get the Grid-IDs from binary matrix

Retrieve the grid ID's from the binary matrix, where the binary code indicates which grid cells are used in the current wind farm constellation.

Usage

get_grids(trimtonOut, Grid)

Arguments

trimtonOut

Input matrix with binary values

Grid

Grid of the considered area

Value

Returns a list of all individuals with X and Y coordinates and the grid cell ID.

See also

Examples

# \donttest{
## Create a random rectangular shapefile
library(sf)
Polygon1 <- sf::st_as_sf(sf::st_sfc(
  sf::st_polygon(list(cbind(
    c(0, 0, 2000, 2000, 0),
    c(0, 2000, 2000, 0, 0)
  ))),
  crs = 3035
))

## Calculate a Grid and an indexed data.frame with coordinates and
## grid cell Ids.
Grid1 <- grid_area(shape = Polygon1, size = 200, prop = 1)
Grid <- Grid1[[1]]
AmountGrids <- nrow(Grid)

startsel <- init_population(Grid, 10, 20)
wind <- data.frame(ws = 12, wd = 0)
wind <- list(wind, probab = 100)
fit <- fitness(
  selection = startsel, referenceHeight = 100, RotorHeight = 100,
  SurfaceRoughness = 0.3, Polygon = Polygon1, resol1 = 200, rot = 20,
  dirspeed = wind, srtm_crop = "", topograp = FALSE, cclRaster = ""
)
allparks <- do.call("rbind", fit)

## SELECTION
## print the amount of Individuals selected.
## Check if the amount of Turbines is as requested.
selec6best <- selection(fit, Grid, 2, TRUE, 6, "VAR")

## CROSSOVER
## u determines the amount of crossover points,
## crossPart determines the method used (Equal/Random),
## uplimit is the maximum allowed permutations
crossOut <- crossover(selec6best, 2, uplimit = 300, crossPart = "RAN")

## MUTATION
## Variable Mutation Rate is activated if more than 2 individuals represent
## the current best solution.
mut <- mutation(a = crossOut, p = 0.3)

## TRIMTON
## After Crossover and Mutation, the amount of turbines in a windpark change
## and have to be corrected to the required amount of turbines.
mut1 <- trimton(
  mut = mut, nturb = 10, allparks = allparks,
  nGrids = AmountGrids, trimForce = FALSE
)

## Get the new Grid-Ids and run a new fitness run.
getRectV <- get_grids(mut1, Grid)
fit <- fitness(
  selection = getRectV, referenceHeight = 100, RotorHeight = 100,
  SurfaceRoughness = 0.3, Polygon = Polygon1, resol1 = 200, rot = 20,
  dirspeed = wind, srtm_crop = "", topograp = FALSE, cclRaster = ""
)
head(fit)
#> $`5,10,17,29,35,37,61,87,96,99`
#>          X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#>  [1,]  900  100     94.0659      7419.022         100   1     20       5
#>  [2,] 1900  100     94.0659      7419.022           0   1     20      10
#>  [3,] 1300  300     94.0659      7419.022         200   1     20      17
#>  [4,] 1700  500     94.0659      7419.022         100   1     20      29
#>  [5,]  900  700     94.0659      7419.022           0   1     20      35
#>  [6,] 1300  700     94.0659      7419.022         100   1     20      37
#>  [7,]  100 1300     94.0659      7419.022           0   1     20      61
#>  [8,] 1300 1700     94.0659      7419.022           0   1     20      87
#>  [9,] 1100 1900     94.0659      7419.022           0   1     20      96
#> [10,] 1700 1900     94.0659      7419.022           0   1     20      99
#>       Parkfitness
#>  [1,]    7419.022
#>  [2,]    7419.022
#>  [3,]    7419.022
#>  [4,]    7419.022
#>  [5,]    7419.022
#>  [6,]    7419.022
#>  [7,]    7419.022
#>  [8,]    7419.022
#>  [9,]    7419.022
#> [10,]    7419.022
#> 
#> $`8,19,20,39,43,49,66,70,92,99`
#>          X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#>  [1,] 1500  100    89.92106      7092.116           0   2     20       8
#>  [2,] 1700  300    89.92106      7092.116         300   2     20      19
#>  [3,] 1900  300    89.92106      7092.116         100   2     20      20
#>  [4,] 1700  700    89.92106      7092.116         200   2     20      39
#>  [5,]  500  900    89.92106      7092.116           0   2     20      43
#>  [6,] 1700  900    89.92106      7092.116         100   2     20      49
#>  [7,] 1100 1300    89.92106      7092.116           0   2     20      66
#>  [8,] 1900 1300    89.92106      7092.116           0   2     20      70
#>  [9,]  300 1900    89.92106      7092.116           0   2     20      92
#> [10,] 1700 1900    89.92106      7092.116           0   2     20      99
#>       Parkfitness
#>  [1,]    7092.116
#>  [2,]    7092.116
#>  [3,]    7092.116
#>  [4,]    7092.116
#>  [5,]    7092.116
#>  [6,]    7092.116
#>  [7,]    7092.116
#>  [8,]    7092.116
#>  [9,]    7092.116
#> [10,]    7092.116
#> 
#> $`7,8,20,27,34,36,50,53,69,99`
#>          X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#>  [1,] 1300  100     93.6917      7389.509         100   3     20       7
#>  [2,] 1500  100     93.6917      7389.509           0   3     20       8
#>  [3,] 1900  300     93.6917      7389.509         100   3     20      20
#>  [4,] 1300  500     93.6917      7389.509           0   3     20      27
#>  [5,]  700  700     93.6917      7389.509           0   3     20      34
#>  [6,] 1100  700     93.6917      7389.509           0   3     20      36
#>  [7,] 1900  900     93.6917      7389.509           0   3     20      50
#>  [8,]  500 1100     93.6917      7389.509           0   3     20      53
#>  [9,] 1700 1300     93.6917      7389.509         100   3     20      69
#> [10,] 1700 1900     93.6917      7389.509           0   3     20      99
#>       Parkfitness
#>  [1,]    7389.509
#>  [2,]    7389.509
#>  [3,]    7389.509
#>  [4,]    7389.509
#>  [5,]    7389.509
#>  [6,]    7389.509
#>  [7,]    7389.509
#>  [8,]    7389.509
#>  [9,]    7389.509
#> [10,]    7389.509
#> 
#> $`9,33,35,37,41,47,67,74,81,99`
#>          X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#>  [1,] 1700  100    90.38575      7128.766         100   4     20       9
#>  [2,]  500  700    90.38575      7128.766           0   4     20      33
#>  [3,]  900  700    90.38575      7128.766           0   4     20      35
#>  [4,] 1300  700    90.38575      7128.766         200   4     20      37
#>  [5,]  100  900    90.38575      7128.766         100   4     20      41
#>  [6,] 1300  900    90.38575      7128.766         100   4     20      47
#>  [7,] 1300 1300    90.38575      7128.766           0   4     20      67
#>  [8,]  700 1500    90.38575      7128.766           0   4     20      74
#>  [9,]  100 1700    90.38575      7128.766           0   4     20      81
#> [10,] 1700 1900    90.38575      7128.766           0   4     20      99
#>       Parkfitness
#>  [1,]    7128.766
#>  [2,]    7128.766
#>  [3,]    7128.766
#>  [4,]    7128.766
#>  [5,]    7128.766
#>  [6,]    7128.766
#>  [7,]    7128.766
#>  [8,]    7128.766
#>  [9,]    7128.766
#> [10,]    7128.766
#> 
#> $`27,29,34,35,44,46,55,60,83,93`
#>          X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#>  [1,] 1300  500    86.91278      6854.852           0   5     20      27
#>  [2,] 1700  500    86.91278      6854.852           0   5     20      29
#>  [3,]  700  700    86.91278      6854.852         100   5     20      34
#>  [4,]  900  700    86.91278      6854.852         100   5     20      35
#>  [5,]  700  900    86.91278      6854.852           0   5     20      44
#>  [6,] 1100  900    86.91278      6854.852           0   5     20      46
#>  [7,]  900 1100    86.91278      6854.852           0   5     20      55
#>  [8,] 1900 1100    86.91278      6854.852           0   5     20      60
#>  [9,]  500 1700    86.91278      6854.852         100   5     20      83
#> [10,]  500 1900    86.91278      6854.852           0   5     20      93
#>       Parkfitness
#>  [1,]    6854.852
#>  [2,]    6854.852
#>  [3,]    6854.852
#>  [4,]    6854.852
#>  [5,]    6854.852
#>  [6,]    6854.852
#>  [7,]    6854.852
#>  [8,]    6854.852
#>  [9,]    6854.852
#> [10,]    6854.852
#> 
#> $`3,4,25,45,52,65,69,77,96,99`
#>          X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#>  [1,]  500  100    92.43145      7290.112           0   6     20       3
#>  [2,]  700  100    92.43145      7290.112           0   6     20       4
#>  [3,]  900  500    92.43145      7290.112         200   6     20      25
#>  [4,]  900  900    92.43145      7290.112         100   6     20      45
#>  [5,]  300 1100    92.43145      7290.112           0   6     20      52
#>  [6,]  900 1300    92.43145      7290.112           0   6     20      65
#>  [7,] 1700 1300    92.43145      7290.112         100   6     20      69
#>  [8,] 1300 1500    92.43145      7290.112           0   6     20      77
#>  [9,] 1100 1900    92.43145      7290.112           0   6     20      96
#> [10,] 1700 1900    92.43145      7290.112           0   6     20      99
#>       Parkfitness
#>  [1,]    7290.112
#>  [2,]    7290.112
#>  [3,]    7290.112
#>  [4,]    7290.112
#>  [5,]    7290.112
#>  [6,]    7290.112
#>  [7,]    7290.112
#>  [8,]    7290.112
#>  [9,]    7290.112
#> [10,]    7290.112
#> 
# }