Get the Grid-IDs from binary matrix

Retrieve the grid ID's from the binary matrix, where the binary code indicates which grid cells are used in the current wind farm constellation.

get_grids(trimtonOut, Grid)

Arguments

trimtonOut

Input matrix with binary values

Grid

Grid of the considered area

Value

Returns a list of all individuals with X and Y coordinates and the grid cell ID.

See also

Examples

# \donttest{
## Create a random rectangular shapefile
library(sf)
Polygon1 <- sf::st_as_sf(sf::st_sfc(
 sf::st_polygon(list(cbind(
   c(0, 0, 2000, 2000, 0),
   c(0, 2000, 2000, 0, 0)))),
crs = 3035
))

## Calculate a Grid and an indexed data.frame with coordinates and grid cell Ids.
Grid1 <- grid_area(shape = Polygon1, size = 200, prop = 1)
Grid <- Grid1[[1]]
AmountGrids <- nrow(Grid)

startsel <- init_population(Grid,10,20)
wind <- data.frame(ws = 12, wd = 0)
wind <- list(wind, probab = 100)
fit <- fitness(selection = startsel,referenceHeight = 100, RotorHeight=100,
               SurfaceRoughness=0.3,Polygon = Polygon1, resol1 = 200,rot=20,
               dirspeed = wind, srtm_crop="",topograp=FALSE,cclRaster="")
allparks <- do.call("rbind",fit)

## SELECTION
## print the amount of Individuals selected.
## Check if the amount of Turbines is as requested.
selec6best <- selection(fit, Grid,2, TRUE, 6, "VAR")

## CROSSOVER
## u determines the amount of crossover points,
## crossPart determines the method used (Equal/Random),
## uplimit is the maximum allowed permutations
crossOut <- crossover(selec6best, 2, uplimit = 300, crossPart="RAN")

## MUTATION
## Variable Mutation Rate is activated if more than 2 individuals represent the
## current best solution.
mut <- mutation(a = crossOut, p = 0.3)

## TRIMTON
## After Crossover and Mutation, the amount of turbines in a windpark change
## and have to be corrected to the required amount of turbines.
mut1 <- trimton(mut = mut, nturb = 10, allparks = allparks,
               nGrids = AmountGrids, trimForce=FALSE)

## Get the new Grid-Ids and run a new fitness run.
getRectV <- get_grids(mut1, Grid)
fit <- fitness(selection = getRectV,referenceHeight = 100, RotorHeight=100,
               SurfaceRoughness=0.3,Polygon = Polygon1, resol1 = 200,rot=20,
               dirspeed = wind, srtm_crop="",topograp=FALSE,cclRaster="")
head(fit)
#> $`22,28,34,46,48,73,82,84,93,95`
#>       X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#> 22  300  500    92.88042      7325.523         100   1     20      22
#> 28 1500  500    92.88042      7325.523         100   1     20      28
#> 34  700  700    92.88042      7325.523         100   1     20      34
#> 46 1100  900    92.88042      7325.523           0   1     20      46
#> 48 1500  900    92.88042      7325.523           0   1     20      48
#> 73  500 1500    92.88042      7325.523         100   1     20      73
#> 82  300 1700    92.88042      7325.523           0   1     20      82
#> 84  700 1700    92.88042      7325.523           0   1     20      84
#> 93  500 1900    92.88042      7325.523           0   1     20      93
#> 95  900 1900    92.88042      7325.523           0   1     20      95
#>    Parkfitness
#> 22    7325.523
#> 28    7325.523
#> 34    7325.523
#> 46    7325.523
#> 48    7325.523
#> 73    7325.523
#> 82    7325.523
#> 84    7325.523
#> 93    7325.523
#> 95    7325.523
#> 
#> $`22,31,36,40,59,62,63,68,83,95`
#>       X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#> 22  300  500     96.0033      7571.826         100   2     20      22
#> 31  100  700     96.0033      7571.826           0   2     20      31
#> 36 1100  700     96.0033      7571.826           0   2     20      36
#> 40 1900  700     96.0033      7571.826           0   2     20      40
#> 59 1700 1100     96.0033      7571.826           0   2     20      59
#> 62  300 1300     96.0033      7571.826           0   2     20      62
#> 63  500 1300     96.0033      7571.826         100   2     20      63
#> 68 1500 1300     96.0033      7571.826           0   2     20      68
#> 83  500 1700     96.0033      7571.826           0   2     20      83
#> 95  900 1900     96.0033      7571.826           0   2     20      95
#>    Parkfitness
#> 22    7571.826
#> 31    7571.826
#> 36    7571.826
#> 40    7571.826
#> 59    7571.826
#> 62    7571.826
#> 63    7571.826
#> 68    7571.826
#> 83    7571.826
#> 95    7571.826
#> 
#> $`5,16,26,41,51,52,58,62,68,92`
#>       X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#> 5   900  100    77.65134      6124.398           0   3     20       5
#> 16 1100  300    77.65134      6124.398         100   3     20      16
#> 26 1100  500    77.65134      6124.398           0   3     20      26
#> 41  100  900    77.65134      6124.398         100   3     20      41
#> 51  100 1100    77.65134      6124.398           0   3     20      51
#> 52  300 1100    77.65134      6124.398         200   3     20      52
#> 58 1500 1100    77.65134      6124.398         100   3     20      58
#> 62  300 1300    77.65134      6124.398         100   3     20      62
#> 68 1500 1300    77.65134      6124.398           0   3     20      68
#> 92  300 1900    77.65134      6124.398           0   3     20      92
#>    Parkfitness
#> 5     6124.398
#> 16    6124.398
#> 26    6124.398
#> 41    6124.398
#> 51    6124.398
#> 52    6124.398
#> 58    6124.398
#> 62    6124.398
#> 68    6124.398
#> 92    6124.398
#> 
#> $`12,24,26,38,69,74,89,93,96,100`
#>        X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#> 12   300  300    95.84169       7559.08           0   4     20      12
#> 24   700  500    95.84169       7559.08         100   4     20      24
#> 26  1100  500    95.84169       7559.08         100   4     20      26
#> 38  1500  700    95.84169       7559.08           0   4     20      38
#> 69  1700 1300    95.84169       7559.08         100   4     20      69
#> 74   700 1500    95.84169       7559.08           0   4     20      74
#> 89  1700 1700    95.84169       7559.08           0   4     20      89
#> 93   500 1900    95.84169       7559.08           0   4     20      93
#> 96  1100 1900    95.84169       7559.08           0   4     20      96
#> 100 1900 1900    95.84169       7559.08           0   4     20     100
#>     Parkfitness
#> 12      7559.08
#> 24      7559.08
#> 26      7559.08
#> 38      7559.08
#> 69      7559.08
#> 74      7559.08
#> 89      7559.08
#> 93      7559.08
#> 96      7559.08
#> 100     7559.08
#> 
#> $`4,24,27,29,34,38,40,48,73,97`
#>       X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#> 4   700  100    86.01947      6784.396         200   5     20       4
#> 24  700  500    86.01947      6784.396         100   5     20      24
#> 27 1300  500    86.01947      6784.396         100   5     20      27
#> 29 1700  500    86.01947      6784.396           0   5     20      29
#> 34  700  700    86.01947      6784.396           0   5     20      34
#> 38 1500  700    86.01947      6784.396         100   5     20      38
#> 40 1900  700    86.01947      6784.396           0   5     20      40
#> 48 1500  900    86.01947      6784.396           0   5     20      48
#> 73  500 1500    86.01947      6784.396           0   5     20      73
#> 97 1300 1900    86.01947      6784.396           0   5     20      97
#>    Parkfitness
#> 4     6784.396
#> 24    6784.396
#> 27    6784.396
#> 29    6784.396
#> 34    6784.396
#> 38    6784.396
#> 40    6784.396
#> 48    6784.396
#> 73    6784.396
#> 97    6784.396
#> 
#> $`5,26,27,28,29,35,80,82,96,98`
#>       X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#> 5   900  100    97.26894      7671.647         100   6     20       5
#> 26 1100  500    97.26894      7671.647         100   6     20      26
#> 27 1300  500    97.26894      7671.647           0   6     20      27
#> 28 1500  500    97.26894      7671.647         100   6     20      28
#> 29 1700  500    97.26894      7671.647           0   6     20      29
#> 35  900  700    97.26894      7671.647           0   6     20      35
#> 80 1900 1500    97.26894      7671.647           0   6     20      80
#> 82  300 1700    97.26894      7671.647           0   6     20      82
#> 96 1100 1900    97.26894      7671.647           0   6     20      96
#> 98 1500 1900    97.26894      7671.647           0   6     20      98
#>    Parkfitness
#> 5     7671.647
#> 26    7671.647
#> 27    7671.647
#> 28    7671.647
#> 29    7671.647
#> 35    7671.647
#> 80    7671.647
#> 82    7671.647
#> 96    7671.647
#> 98    7671.647
#> 

# }