Get the Grid-IDs from binary matrix

Retrieve the grid ID's from the binary matrix, where the binary code indicates which grid cells are used in the current wind farm constellation.

get_grids(trimtonOut, Grid)

Arguments

trimtonOut

Input matrix with binary values

Grid

Grid of the considered area

Value

Returns a list of all individuals with X and Y coordinates and the grid cell ID.

See also

Examples

# \donttest{
## Create a random rectangular shapefile
library(sf)
Polygon1 <- sf::st_as_sf(sf::st_sfc(
 sf::st_polygon(list(cbind(
   c(0, 0, 2000, 2000, 0),
   c(0, 2000, 2000, 0, 0)))),
crs = 3035
))

## Calculate a Grid and an indexed data.frame with coordinates and grid cell Ids.
Grid1 <- grid_area(shape = Polygon1, size = 200, prop = 1)
Grid <- Grid1[[1]]
AmountGrids <- nrow(Grid)

startsel <- init_population(Grid,10,20)
wind <- data.frame(ws = 12, wd = 0)
wind <- list(wind, probab = 100)
fit <- fitness(selection = startsel,referenceHeight = 100, RotorHeight=100,
               SurfaceRoughness=0.3,Polygon = Polygon1, resol1 = 200,rot=20,
               dirspeed = wind, srtm_crop="",topograp=FALSE,cclRaster="")
allparks <- do.call("rbind",fit)

## SELECTION
## print the amount of Individuals selected.
## Check if the amount of Turbines is as requested.
selec6best <- selection(fit, Grid,2, TRUE, 6, "VAR")

## CROSSOVER
## u determines the amount of crossover points,
## crossPart determines the method used (Equal/Random),
## uplimit is the maximum allowed permutations
crossOut <- crossover(selec6best, 2, uplimit = 300, crossPart="RAN")

## MUTATION
## Variable Mutation Rate is activated if more than 2 individuals represent the
## current best solution.
mut <- mutation(a = crossOut, p = 0.3)

## TRIMTON
## After Crossover and Mutation, the amount of turbines in a windpark change
## and have to be corrected to the required amount of turbines.
mut1 <- trimton(mut = mut, nturb = 10, allparks = allparks,
               nGrids = AmountGrids, trimForce=FALSE)

## Get the new Grid-Ids and run a new fitness run.
getRectV <- get_grids(mut1, Grid)
fit <- fitness(selection = getRectV,referenceHeight = 100, RotorHeight=100,
               SurfaceRoughness=0.3,Polygon = Polygon1, resol1 = 200,rot=20,
               dirspeed = wind, srtm_crop="",topograp=FALSE,cclRaster="")
head(fit)
#> $`4,17,26,64,78,80,82,91,98,99`
#>          X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#>  [1,]  700  100    96.54512      7614.559         100   1     20       4
#>  [2,] 1300  300    96.54512      7614.559           0   1     20      17
#>  [3,] 1100  500    96.54512      7614.559           0   1     20      26
#>  [4,]  700 1300    96.54512      7614.559           0   1     20      64
#>  [5,] 1500 1500    96.54512      7614.559         100   1     20      78
#>  [6,] 1900 1500    96.54512      7614.559           0   1     20      80
#>  [7,]  300 1700    96.54512      7614.559           0   1     20      82
#>  [8,]  100 1900    96.54512      7614.559           0   1     20      91
#>  [9,] 1500 1900    96.54512      7614.559           0   1     20      98
#> [10,] 1700 1900    96.54512      7614.559           0   1     20      99
#>       Parkfitness
#>  [1,]    7614.559
#>  [2,]    7614.559
#>  [3,]    7614.559
#>  [4,]    7614.559
#>  [5,]    7614.559
#>  [6,]    7614.559
#>  [7,]    7614.559
#>  [8,]    7614.559
#>  [9,]    7614.559
#> [10,]    7614.559
#> 
#> $`30,33,35,38,46,52,61,62,97,98`
#>          X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#>  [1,] 1900  500     94.2324      7432.154           0   2     20      30
#>  [2,]  500  700     94.2324      7432.154           0   2     20      33
#>  [3,]  900  700     94.2324      7432.154           0   2     20      35
#>  [4,] 1500  700     94.2324      7432.154         100   2     20      38
#>  [5,] 1100  900     94.2324      7432.154           0   2     20      46
#>  [6,]  300 1100     94.2324      7432.154         100   2     20      52
#>  [7,]  100 1300     94.2324      7432.154           0   2     20      61
#>  [8,]  300 1300     94.2324      7432.154           0   2     20      62
#>  [9,] 1300 1900     94.2324      7432.154           0   2     20      97
#> [10,] 1500 1900     94.2324      7432.154           0   2     20      98
#>       Parkfitness
#>  [1,]    7432.154
#>  [2,]    7432.154
#>  [3,]    7432.154
#>  [4,]    7432.154
#>  [5,]    7432.154
#>  [6,]    7432.154
#>  [7,]    7432.154
#>  [8,]    7432.154
#>  [9,]    7432.154
#> [10,]    7432.154
#> 
#> $`3,23,24,28,35,39,40,46,64,98`
#>          X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#>  [1,]  500  100     95.5097      7532.895         100   3     20       3
#>  [2,]  500  500     95.5097      7532.895           0   3     20      23
#>  [3,]  700  500     95.5097      7532.895         100   3     20      24
#>  [4,] 1500  500     95.5097      7532.895         100   3     20      28
#>  [5,]  900  700     95.5097      7532.895           0   3     20      35
#>  [6,] 1700  700     95.5097      7532.895           0   3     20      39
#>  [7,] 1900  700     95.5097      7532.895           0   3     20      40
#>  [8,] 1100  900     95.5097      7532.895           0   3     20      46
#>  [9,]  700 1300     95.5097      7532.895           0   3     20      64
#> [10,] 1500 1900     95.5097      7532.895           0   3     20      98
#>       Parkfitness
#>  [1,]    7532.895
#>  [2,]    7532.895
#>  [3,]    7532.895
#>  [4,]    7532.895
#>  [5,]    7532.895
#>  [6,]    7532.895
#>  [7,]    7532.895
#>  [8,]    7532.895
#>  [9,]    7532.895
#> [10,]    7532.895
#> 
#> $`11,30,31,33,36,37,75,76,86,93`
#>          X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#>  [1,]  100  300    89.98345      7097.037         100   4     20      11
#>  [2,] 1900  500    89.98345      7097.037           0   4     20      30
#>  [3,]  100  700    89.98345      7097.037           0   4     20      31
#>  [4,]  500  700    89.98345      7097.037         100   4     20      33
#>  [5,] 1100  700    89.98345      7097.037         200   4     20      36
#>  [6,] 1300  700    89.98345      7097.037           0   4     20      37
#>  [7,]  900 1500    89.98345      7097.037           0   4     20      75
#>  [8,] 1100 1500    89.98345      7097.037         100   4     20      76
#>  [9,] 1100 1700    89.98345      7097.037           0   4     20      86
#> [10,]  500 1900    89.98345      7097.037           0   4     20      93
#>       Parkfitness
#>  [1,]    7097.037
#>  [2,]    7097.037
#>  [3,]    7097.037
#>  [4,]    7097.037
#>  [5,]    7097.037
#>  [6,]    7097.037
#>  [7,]    7097.037
#>  [8,]    7097.037
#>  [9,]    7097.037
#> [10,]    7097.037
#> 
#> $`9,10,21,33,51,57,64,70,77,98`
#>          X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#>  [1,] 1700  100    94.80127      7477.021           0   5     20       9
#>  [2,] 1900  100    94.80127      7477.021         100   5     20      10
#>  [3,]  100  500    94.80127      7477.021         100   5     20      21
#>  [4,]  500  700    94.80127      7477.021           0   5     20      33
#>  [5,]  100 1100    94.80127      7477.021           0   5     20      51
#>  [6,] 1300 1100    94.80127      7477.021         100   5     20      57
#>  [7,]  700 1300    94.80127      7477.021           0   5     20      64
#>  [8,] 1900 1300    94.80127      7477.021           0   5     20      70
#>  [9,] 1300 1500    94.80127      7477.021           0   5     20      77
#> [10,] 1500 1900    94.80127      7477.021           0   5     20      98
#>       Parkfitness
#>  [1,]    7477.021
#>  [2,]    7477.021
#>  [3,]    7477.021
#>  [4,]    7477.021
#>  [5,]    7477.021
#>  [6,]    7477.021
#>  [7,]    7477.021
#>  [8,]    7477.021
#>  [9,]    7477.021
#> [10,]    7477.021
#> 
#> $`21,23,46,53,54,83,86,92,96,98`
#>          X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#>  [1,]  100  500    89.85427      7086.849           0   6     20      21
#>  [2,]  500  500    89.85427      7086.849         200   6     20      23
#>  [3,] 1100  900    89.85427      7086.849         200   6     20      46
#>  [4,]  500 1100    89.85427      7086.849         100   6     20      53
#>  [5,]  700 1100    89.85427      7086.849           0   6     20      54
#>  [6,]  500 1700    89.85427      7086.849           0   6     20      83
#>  [7,] 1100 1700    89.85427      7086.849         100   6     20      86
#>  [8,]  300 1900    89.85427      7086.849           0   6     20      92
#>  [9,] 1100 1900    89.85427      7086.849           0   6     20      96
#> [10,] 1500 1900    89.85427      7086.849           0   6     20      98
#>       Parkfitness
#>  [1,]    7086.849
#>  [2,]    7086.849
#>  [3,]    7086.849
#>  [4,]    7086.849
#>  [5,]    7086.849
#>  [6,]    7086.849
#>  [7,]    7086.849
#>  [8,]    7086.849
#>  [9,]    7086.849
#> [10,]    7086.849
#> 

# }