Get the Grid-IDs from binary matrix

Retrieve the grid ID's from the binary matrix, where the binary code indicates which grid cells are used in the current wind farm constellation.

Usage

get_grids(trimtonOut, Grid)

Arguments

trimtonOut

Input matrix with binary values

Grid

Grid of the considered area

Value

Returns a list of all individuals with X and Y coordinates and the grid cell ID.

See also

Examples

# \donttest{
## Create a random rectangular shapefile
library(sf)
Polygon1 <- sf::st_as_sf(sf::st_sfc(
 sf::st_polygon(list(cbind(
   c(0, 0, 2000, 2000, 0),
   c(0, 2000, 2000, 0, 0)))),
crs = 3035
))

## Calculate a Grid and an indexed data.frame with coordinates and grid cell Ids.
Grid1 <- grid_area(shape = Polygon1, size = 200, prop = 1)
Grid <- Grid1[[1]]
AmountGrids <- nrow(Grid)

startsel <- init_population(Grid,10,20)
wind <- data.frame(ws = 12, wd = 0)
wind <- list(wind, probab = 100)
fit <- fitness(selection = startsel,referenceHeight = 100, RotorHeight=100,
               SurfaceRoughness=0.3,Polygon = Polygon1, resol1 = 200,rot=20,
               dirspeed = wind, srtm_crop="",topograp=FALSE,cclRaster="")
allparks <- do.call("rbind",fit)

## SELECTION
## print the amount of Individuals selected.
## Check if the amount of Turbines is as requested.
selec6best <- selection(fit, Grid,2, TRUE, 6, "VAR")

## CROSSOVER
## u determines the amount of crossover points,
## crossPart determines the method used (Equal/Random),
## uplimit is the maximum allowed permutations
crossOut <- crossover(selec6best, 2, uplimit = 300, crossPart="RAN")

## MUTATION
## Variable Mutation Rate is activated if more than 2 individuals represent the
## current best solution.
mut <- mutation(a = crossOut, p = 0.3)

## TRIMTON
## After Crossover and Mutation, the amount of turbines in a windpark change
## and have to be corrected to the required amount of turbines.
mut1 <- trimton(mut = mut, nturb = 10, allparks = allparks,
               nGrids = AmountGrids, trimForce=FALSE)

## Get the new Grid-Ids and run a new fitness run.
getRectV <- get_grids(mut1, Grid)
fit <- fitness(selection = getRectV,referenceHeight = 100, RotorHeight=100,
               SurfaceRoughness=0.3,Polygon = Polygon1, resol1 = 200,rot=20,
               dirspeed = wind, srtm_crop="",topograp=FALSE,cclRaster="")
head(fit)
#> $`2,7,15,26,59,63,80,81,84,85`
#>          X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#>  [1,]  300  100    99.50639      7848.116           0   1     20       2
#>  [2,] 1300  100    99.50639      7848.116           0   1     20       7
#>  [3,]  900  300    99.50639      7848.116         100   1     20      15
#>  [4,] 1100  500    99.50639      7848.116           0   1     20      26
#>  [5,] 1700 1100    99.50639      7848.116           0   1     20      59
#>  [6,]  500 1300    99.50639      7848.116           0   1     20      63
#>  [7,] 1900 1500    99.50639      7848.116           0   1     20      80
#>  [8,]  100 1700    99.50639      7848.116           0   1     20      81
#>  [9,]  700 1700    99.50639      7848.116           0   1     20      84
#> [10,]  900 1700    99.50639      7848.116           0   1     20      85
#>       Parkfitness
#>  [1,]    7848.116
#>  [2,]    7848.116
#>  [3,]    7848.116
#>  [4,]    7848.116
#>  [5,]    7848.116
#>  [6,]    7848.116
#>  [7,]    7848.116
#>  [8,]    7848.116
#>  [9,]    7848.116
#> [10,]    7848.116
#> 
#> $`6,23,24,25,34,68,78,84,85,92`
#>          X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#>  [1,] 1100  100    88.23476      6959.117           0   2     20       6
#>  [2,]  500  500    88.23476      6959.117           0   2     20      23
#>  [3,]  700  500    88.23476      6959.117         200   2     20      24
#>  [4,]  900  500    88.23476      6959.117         100   2     20      25
#>  [5,]  700  700    88.23476      6959.117         100   2     20      34
#>  [6,] 1500 1300    88.23476      6959.117         100   2     20      68
#>  [7,] 1500 1500    88.23476      6959.117           0   2     20      78
#>  [8,]  700 1700    88.23476      6959.117           0   2     20      84
#>  [9,]  900 1700    88.23476      6959.117           0   2     20      85
#> [10,]  300 1900    88.23476      6959.117           0   2     20      92
#>       Parkfitness
#>  [1,]    6959.117
#>  [2,]    6959.117
#>  [3,]    6959.117
#>  [4,]    6959.117
#>  [5,]    6959.117
#>  [6,]    6959.117
#>  [7,]    6959.117
#>  [8,]    6959.117
#>  [9,]    6959.117
#> [10,]    6959.117
#> 
#> $`13,21,29,37,43,46,59,69,70,76`
#>          X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#>  [1,]  500  300    89.31546      7044.353         100   3     20      13
#>  [2,]  100  500    89.31546      7044.353           0   3     20      21
#>  [3,] 1700  500    89.31546      7044.353         200   3     20      29
#>  [4,] 1300  700    89.31546      7044.353           0   3     20      37
#>  [5,]  500  900    89.31546      7044.353           0   3     20      43
#>  [6,] 1100  900    89.31546      7044.353         100   3     20      46
#>  [7,] 1700 1100    89.31546      7044.353         100   3     20      59
#>  [8,] 1700 1300    89.31546      7044.353           0   3     20      69
#>  [9,] 1900 1300    89.31546      7044.353           0   3     20      70
#> [10,] 1100 1500    89.31546      7044.353           0   3     20      76
#>       Parkfitness
#>  [1,]    7044.353
#>  [2,]    7044.353
#>  [3,]    7044.353
#>  [4,]    7044.353
#>  [5,]    7044.353
#>  [6,]    7044.353
#>  [7,]    7044.353
#>  [8,]    7044.353
#>  [9,]    7044.353
#> [10,]    7044.353
#> 
#> $`1,15,16,28,42,47,53,73,89,93`
#>          X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#>  [1,]  100  100     94.1753       7427.65           0   4     20       1
#>  [2,]  900  300     94.1753       7427.65           0   4     20      15
#>  [3,] 1100  300     94.1753       7427.65           0   4     20      16
#>  [4,] 1500  500     94.1753       7427.65           0   4     20      28
#>  [5,]  300  900     94.1753       7427.65           0   4     20      42
#>  [6,] 1300  900     94.1753       7427.65           0   4     20      47
#>  [7,]  500 1100     94.1753       7427.65         200   4     20      53
#>  [8,]  500 1500     94.1753       7427.65         100   4     20      73
#>  [9,] 1700 1700     94.1753       7427.65           0   4     20      89
#> [10,]  500 1900     94.1753       7427.65           0   4     20      93
#>       Parkfitness
#>  [1,]     7427.65
#>  [2,]     7427.65
#>  [3,]     7427.65
#>  [4,]     7427.65
#>  [5,]     7427.65
#>  [6,]     7427.65
#>  [7,]     7427.65
#>  [8,]     7427.65
#>  [9,]     7427.65
#> [10,]     7427.65
#> 
#> $`5,29,37,38,71,73,74,77,91,93`
#>          X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#>  [1,]  900  100    93.18271      7349.364           0   5     20       5
#>  [2,] 1700  500    93.18271      7349.364           0   5     20      29
#>  [3,] 1300  700    93.18271      7349.364         100   5     20      37
#>  [4,] 1500  700    93.18271      7349.364           0   5     20      38
#>  [5,]  100 1500    93.18271      7349.364         100   5     20      71
#>  [6,]  500 1500    93.18271      7349.364         100   5     20      73
#>  [7,]  700 1500    93.18271      7349.364           0   5     20      74
#>  [8,] 1300 1500    93.18271      7349.364           0   5     20      77
#>  [9,]  100 1900    93.18271      7349.364           0   5     20      91
#> [10,]  500 1900    93.18271      7349.364           0   5     20      93
#>       Parkfitness
#>  [1,]    7349.364
#>  [2,]    7349.364
#>  [3,]    7349.364
#>  [4,]    7349.364
#>  [5,]    7349.364
#>  [6,]    7349.364
#>  [7,]    7349.364
#>  [8,]    7349.364
#>  [9,]    7349.364
#> [10,]    7349.364
#> 
#> $`12,19,27,31,37,43,49,79,90,91`
#>          X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#>  [1,]  300  300    90.64835      7149.478           0   6     20      12
#>  [2,] 1700  300    90.64835      7149.478         200   6     20      19
#>  [3,] 1300  500    90.64835      7149.478         100   6     20      27
#>  [4,]  100  700    90.64835      7149.478         100   6     20      31
#>  [5,] 1300  700    90.64835      7149.478           0   6     20      37
#>  [6,]  500  900    90.64835      7149.478           0   6     20      43
#>  [7,] 1700  900    90.64835      7149.478         100   6     20      49
#>  [8,] 1700 1500    90.64835      7149.478           0   6     20      79
#>  [9,] 1900 1700    90.64835      7149.478           0   6     20      90
#> [10,]  100 1900    90.64835      7149.478           0   6     20      91
#>       Parkfitness
#>  [1,]    7149.478
#>  [2,]    7149.478
#>  [3,]    7149.478
#>  [4,]    7149.478
#>  [5,]    7149.478
#>  [6,]    7149.478
#>  [7,]    7149.478
#>  [8,]    7149.478
#>  [9,]    7149.478
#> [10,]    7149.478
#> 

# }