Get the Grid-IDs from binary matrix

Get the grid IDs from the trimmed binary matrix, where the binary code indicates which grid cells are used in the current wind farm constellation.

get_grids(trimtonOut, Grid)

Arguments

trimtonOut
Input matrix with binary values.

Grid
Grid of the considered area

Value

Returns a list of all individuals with X and Y coordinates and the grid cell ID.

See also

Examples

## Create a random rectangular shapefile
library(sp)
Polygon1 <- Polygon(rbind(c(0, 0), c(0, 2000), c(2000, 2000), c(2000, 0)))
Polygon1 <- Polygons(list(Polygon1),1);
Polygon1 <- SpatialPolygons(list(Polygon1))
Projection <- "+proj=laea +lat_0=52 +lon_0=10 +x_0=4321000 +y_0=3210000
+ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs"
proj4string(Polygon1) <- CRS(Projection)

## Calculate a Grid and an indexed data.frame with coordinates and grid cell Ids.
Grid1 <- grid_area(shape = Polygon1,resol = 200,prop = 1);
Grid <- Grid1[[1]]
AmountGrids <- nrow(Grid)

startsel <- init_population(Grid,10,20);
wind <- data.frame(ws = 12, wd = 0)
wind <- list(wind, probab = 100)
fit <- fitness(selection = startsel,referenceHeight = 100, RotorHeight=100,
               SurfaceRoughness=0.3,Polygon = Polygon1, resol1 = 200,rot=20,
               dirspeed = wind, srtm_crop="",topograp=FALSE,cclRaster="")
allparks <- do.call("rbind",fit);

## SELECTION
## print the amount of Individuals selected.
## Check if the amount of Turbines is as requested.
selec6best <- selection(fit, Grid,2, TRUE, 6, "VAR");

## CROSSOVER
## u determines the amount of crossover points,
## crossPart determines the method used (Equal/Random),
## uplimit is the maximum allowed permutations
crossOut <- crossover(selec6best, 2, uplimit = 300, crossPart="RAN");

## MUTATION
## Variable Mutation Rate is activated if more than 2 individuals represent the
## current best solution.
mut <- mutation(a = crossOut, p = 0.3);

## TRIMTON
## After Crossover and Mutation, the amount of turbines in a windpark change
## and have to be corrected to the required amount of turbines.
mut1 <- trimton(mut = mut, nturb = 10, allparks = allparks,
               nGrids = AmountGrids, trimForce=FALSE)

## Get the new Grid-Ids and run a new fitness run.
getRectV <- get_grids(mut1, Grid)
fit <- fitness(selection = getRectV,referenceHeight = 100, RotorHeight=100,
               SurfaceRoughness=0.3,Polygon = Polygon1, resol1 = 200,rot=20,
               dirspeed = wind, srtm_crop="",topograp=FALSE,cclRaster="")
head(fit)

#> $`21,22,30,32,67,77,78,80,81,82`
#>       X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#> 21  100 1500    87.20954      6878.258           0   1     20      21
#> 22  300 1500    87.20954      6878.258           0   1     20      22
#> 30 1900 1500    87.20954      6878.258           0   1     20      30
#> 32  300 1300    87.20954      6878.258         100   1     20      32
#> 67 1300  700    87.20954      6878.258           0   1     20      67
#> 77 1300  500    87.20954      6878.258         100   1     20      77
#> 78 1500  500    87.20954      6878.258           0   1     20      78
#> 80 1900  500    87.20954      6878.258         100   1     20      80
#> 81  100  300    87.20954      6878.258         100   1     20      81
#> 82  300  300    87.20954      6878.258         200   1     20      82
#>    Parkfitness
#> 21    6878.258
#> 22    6878.258
#> 30    6878.258
#> 32    6878.258
#> 67    6878.258
#> 77    6878.258
#> 78    6878.258
#> 80    6878.258
#> 81    6878.258
#> 82    6878.258
#> 
#> $`13,20,38,42,46,49,53,70,81,99`
#>       X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#> 13  500 1700    97.13568      7661.137           0   2     20      13
#> 20 1900 1700    97.13568      7661.137           0   2     20      20
#> 38 1500 1300    97.13568      7661.137           0   2     20      38
#> 42  300 1100    97.13568      7661.137           0   2     20      42
#> 46 1100 1100    97.13568      7661.137           0   2     20      46
#> 49 1700 1100    97.13568      7661.137           0   2     20      49
#> 53  500  900    97.13568      7661.137         100   2     20      53
#> 70 1900  700    97.13568      7661.137         100   2     20      70
#> 81  100  300    97.13568      7661.137           0   2     20      81
#> 99 1700  100    97.13568      7661.137         100   2     20      99
#>    Parkfitness
#> 13    7661.137
#> 20    7661.137
#> 38    7661.137
#> 42    7661.137
#> 46    7661.137
#> 49    7661.137
#> 53    7661.137
#> 70    7661.137
#> 81    7661.137
#> 99    7661.137
#> 
#> $`1,20,36,40,51,59,78,81,84,93`
#>       X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#> 1   100 1900    94.55398      7457.517           0   3     20       1
#> 20 1900 1700    94.55398      7457.517           0   3     20      20
#> 36 1100 1300    94.55398      7457.517           0   3     20      36
#> 40 1900 1300    94.55398      7457.517         100   3     20      40
#> 51  100  900    94.55398      7457.517         100   3     20      51
#> 59 1700  900    94.55398      7457.517           0   3     20      59
#> 78 1500  500    94.55398      7457.517           0   3     20      78
#> 81  100  300    94.55398      7457.517         200   3     20      81
#> 84  700  300    94.55398      7457.517           0   3     20      84
#> 93  500  100    94.55398      7457.517           0   3     20      93
#>    Parkfitness
#> 1     7457.517
#> 20    7457.517
#> 36    7457.517
#> 40    7457.517
#> 51    7457.517
#> 59    7457.517
#> 78    7457.517
#> 81    7457.517
#> 84    7457.517
#> 93    7457.517
#> 
#> $`12,14,17,31,33,55,56,59,72,92`
#>       X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#> 12  300 1700    96.52476      7612.954           0   4     20      12
#> 14  700 1700    96.52476      7612.954           0   4     20      14
#> 17 1300 1700    96.52476      7612.954           0   4     20      17
#> 31  100 1300    96.52476      7612.954           0   4     20      31
#> 33  500 1300    96.52476      7612.954           0   4     20      33
#> 55  900  900    96.52476      7612.954           0   4     20      55
#> 56 1100  900    96.52476      7612.954           0   4     20      56
#> 59 1700  900    96.52476      7612.954           0   4     20      59
#> 72  300  500    96.52476      7612.954         100   4     20      72
#> 92  300  100    96.52476      7612.954         200   4     20      92
#>    Parkfitness
#> 12    7612.954
#> 14    7612.954
#> 17    7612.954
#> 31    7612.954
#> 33    7612.954
#> 55    7612.954
#> 56    7612.954
#> 59    7612.954
#> 72    7612.954
#> 92    7612.954
#> 
#> $`11,23,29,33,46,47,55,69,75,99`
#>       X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#> 11  100 1700    89.06761      7024.804           0   5     20      11
#> 23  500 1500    89.06761      7024.804           0   5     20      23
#> 29 1700 1500    89.06761      7024.804           0   5     20      29
#> 33  500 1300    89.06761      7024.804         100   5     20      33
#> 46 1100 1100    89.06761      7024.804           0   5     20      46
#> 47 1300 1100    89.06761      7024.804           0   5     20      47
#> 55  900  900    89.06761      7024.804           0   5     20      55
#> 69 1700  700    89.06761      7024.804         100   5     20      69
#> 75  900  500    89.06761      7024.804         100   5     20      75
#> 99 1700  100    89.06761      7024.804         200   5     20      99
#>    Parkfitness
#> 11    7024.804
#> 23    7024.804
#> 29    7024.804
#> 33    7024.804
#> 46    7024.804
#> 47    7024.804
#> 55    7024.804
#> 69    7024.804
#> 75    7024.804
#> 99    7024.804
#> 
#> $`14,34,36,53,59,76,77,78,93,98`
#>       X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#> 14  700 1700    92.00661      7256.605           0   6     20      14
#> 34  700 1300    92.00661      7256.605         100   6     20      34
#> 36 1100 1300    92.00661      7256.605           0   6     20      36
#> 53  500  900    92.00661      7256.605           0   6     20      53
#> 59 1700  900    92.00661      7256.605           0   6     20      59
#> 76 1100  500    92.00661      7256.605         100   6     20      76
#> 77 1300  500    92.00661      7256.605           0   6     20      77
#> 78 1500  500    92.00661      7256.605           0   6     20      78
#> 93  500  100    92.00661      7256.605         100   6     20      93
#> 98 1500  100    92.00661      7256.605         100   6     20      98
#>    Parkfitness
#> 14    7256.605
#> 34    7256.605
#> 36    7256.605
#> 53    7256.605
#> 59    7256.605
#> 76    7256.605
#> 77    7256.605
#> 78    7256.605
#> 93    7256.605
#> 98    7256.605
#>