Get the Grid-IDs from binary matrix

Retrieve the grid ID's from the binary matrix, where the binary code indicates which grid cells are used in the current wind farm constellation.

Usage

get_grids(trimtonOut, Grid)

Arguments

trimtonOut

Input matrix with binary values

Grid

Grid of the considered area

Value

Returns a list of all individuals with X and Y coordinates and the grid cell ID.

See also

Examples

# \donttest{
## Create a random rectangular shapefile
library(sf)
Polygon1 <- sf::st_as_sf(sf::st_sfc(
 sf::st_polygon(list(cbind(
   c(0, 0, 2000, 2000, 0),
   c(0, 2000, 2000, 0, 0)))),
crs = 3035
))

## Calculate a Grid and an indexed data.frame with coordinates and grid cell Ids.
Grid1 <- grid_area(shape = Polygon1, size = 200, prop = 1)
Grid <- Grid1[[1]]
AmountGrids <- nrow(Grid)

startsel <- init_population(Grid,10,20)
wind <- data.frame(ws = 12, wd = 0)
wind <- list(wind, probab = 100)
fit <- fitness(selection = startsel,referenceHeight = 100, RotorHeight=100,
               SurfaceRoughness=0.3,Polygon = Polygon1, resol1 = 200,rot=20,
               dirspeed = wind, srtm_crop="",topograp=FALSE,cclRaster="")
allparks <- do.call("rbind",fit)

## SELECTION
## print the amount of Individuals selected.
## Check if the amount of Turbines is as requested.
selec6best <- selection(fit, Grid,2, TRUE, 6, "VAR")

## CROSSOVER
## u determines the amount of crossover points,
## crossPart determines the method used (Equal/Random),
## uplimit is the maximum allowed permutations
crossOut <- crossover(selec6best, 2, uplimit = 300, crossPart="RAN")

## MUTATION
## Variable Mutation Rate is activated if more than 2 individuals represent the
## current best solution.
mut <- mutation(a = crossOut, p = 0.3)

## TRIMTON
## After Crossover and Mutation, the amount of turbines in a windpark change
## and have to be corrected to the required amount of turbines.
mut1 <- trimton(mut = mut, nturb = 10, allparks = allparks,
               nGrids = AmountGrids, trimForce=FALSE)

## Get the new Grid-Ids and run a new fitness run.
getRectV <- get_grids(mut1, Grid)
fit <- fitness(selection = getRectV,referenceHeight = 100, RotorHeight=100,
               SurfaceRoughness=0.3,Polygon = Polygon1, resol1 = 200,rot=20,
               dirspeed = wind, srtm_crop="",topograp=FALSE,cclRaster="")
head(fit)
#> $`6,33,49,55,58,59,71,81,85,96`
#>          X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#>  [1,] 1100  100    87.66661      6914.307         100   1     20       6
#>  [2,]  500  700    87.66661      6914.307           0   1     20      33
#>  [3,] 1700  900    87.66661      6914.307         100   1     20      49
#>  [4,]  900 1100    87.66661      6914.307         100   1     20      55
#>  [5,] 1500 1100    87.66661      6914.307           0   1     20      58
#>  [6,] 1700 1100    87.66661      6914.307           0   1     20      59
#>  [7,]  100 1500    87.66661      6914.307         100   1     20      71
#>  [8,]  100 1700    87.66661      6914.307           0   1     20      81
#>  [9,]  900 1700    87.66661      6914.307           0   1     20      85
#> [10,] 1100 1900    87.66661      6914.307           0   1     20      96
#>       Parkfitness
#>  [1,]    6914.307
#>  [2,]    6914.307
#>  [3,]    6914.307
#>  [4,]    6914.307
#>  [5,]    6914.307
#>  [6,]    6914.307
#>  [7,]    6914.307
#>  [8,]    6914.307
#>  [9,]    6914.307
#> [10,]    6914.307
#> 
#> $`4,14,24,39,48,49,53,62,70,94`
#>          X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#>  [1,]  700  100    83.65812      6598.156         300   2     20       4
#>  [2,]  700  300    83.65812      6598.156         200   2     20      14
#>  [3,]  700  500    83.65812      6598.156         100   2     20      24
#>  [4,] 1700  700    83.65812      6598.156         100   2     20      39
#>  [5,] 1500  900    83.65812      6598.156           0   2     20      48
#>  [6,] 1700  900    83.65812      6598.156           0   2     20      49
#>  [7,]  500 1100    83.65812      6598.156           0   2     20      53
#>  [8,]  300 1300    83.65812      6598.156           0   2     20      62
#>  [9,] 1900 1300    83.65812      6598.156           0   2     20      70
#> [10,]  700 1900    83.65812      6598.156           0   2     20      94
#>       Parkfitness
#>  [1,]    6598.156
#>  [2,]    6598.156
#>  [3,]    6598.156
#>  [4,]    6598.156
#>  [5,]    6598.156
#>  [6,]    6598.156
#>  [7,]    6598.156
#>  [8,]    6598.156
#>  [9,]    6598.156
#> [10,]    6598.156
#> 
#> $`3,8,33,43,55,59,72,79,81,84`
#>          X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#>  [1,]  500  100    89.98257      7096.968         200   3     20       3
#>  [2,] 1500  100    89.98257      7096.968           0   3     20       8
#>  [3,]  500  700    89.98257      7096.968         100   3     20      33
#>  [4,]  500  900    89.98257      7096.968           0   3     20      43
#>  [5,]  900 1100    89.98257      7096.968           0   3     20      55
#>  [6,] 1700 1100    89.98257      7096.968         100   3     20      59
#>  [7,]  300 1500    89.98257      7096.968           0   3     20      72
#>  [8,] 1700 1500    89.98257      7096.968           0   3     20      79
#>  [9,]  100 1700    89.98257      7096.968           0   3     20      81
#> [10,]  700 1700    89.98257      7096.968           0   3     20      84
#>       Parkfitness
#>  [1,]    7096.968
#>  [2,]    7096.968
#>  [3,]    7096.968
#>  [4,]    7096.968
#>  [5,]    7096.968
#>  [6,]    7096.968
#>  [7,]    7096.968
#>  [8,]    7096.968
#>  [9,]    7096.968
#> [10,]    7096.968
#> 
#> $`2,11,19,38,45,47,61,83,85,88`
#>          X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#>  [1,]  300  100    97.13568      7661.137           0   4     20       2
#>  [2,]  100  300    97.13568      7661.137         100   4     20      11
#>  [3,] 1700  300    97.13568      7661.137           0   4     20      19
#>  [4,] 1500  700    97.13568      7661.137         100   4     20      38
#>  [5,]  900  900    97.13568      7661.137         100   4     20      45
#>  [6,] 1300  900    97.13568      7661.137           0   4     20      47
#>  [7,]  100 1300    97.13568      7661.137           0   4     20      61
#>  [8,]  500 1700    97.13568      7661.137           0   4     20      83
#>  [9,]  900 1700    97.13568      7661.137           0   4     20      85
#> [10,] 1500 1700    97.13568      7661.137           0   4     20      88
#>       Parkfitness
#>  [1,]    7661.137
#>  [2,]    7661.137
#>  [3,]    7661.137
#>  [4,]    7661.137
#>  [5,]    7661.137
#>  [6,]    7661.137
#>  [7,]    7661.137
#>  [8,]    7661.137
#>  [9,]    7661.137
#> [10,]    7661.137
#> 
#> $`1,18,28,31,32,50,52,58,68,94`
#>          X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#>  [1,]  100  100    82.99861      6546.139         100   5     20       1
#>  [2,] 1500  300    82.99861      6546.139         300   5     20      18
#>  [3,] 1500  500    82.99861      6546.139         200   5     20      28
#>  [4,]  100  700    82.99861      6546.139           0   5     20      31
#>  [5,]  300  700    82.99861      6546.139         100   5     20      32
#>  [6,] 1900  900    82.99861      6546.139           0   5     20      50
#>  [7,]  300 1100    82.99861      6546.139           0   5     20      52
#>  [8,] 1500 1100    82.99861      6546.139         100   5     20      58
#>  [9,] 1500 1300    82.99861      6546.139           0   5     20      68
#> [10,]  700 1900    82.99861      6546.139           0   5     20      94
#>       Parkfitness
#>  [1,]    6546.139
#>  [2,]    6546.139
#>  [3,]    6546.139
#>  [4,]    6546.139
#>  [5,]    6546.139
#>  [6,]    6546.139
#>  [7,]    6546.139
#>  [8,]    6546.139
#>  [9,]    6546.139
#> [10,]    6546.139
#> 
#> $`2,6,34,43,49,61,72,87,89,98`
#>          X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#>  [1,]  300  100    98.33029      7755.357         100   6     20       2
#>  [2,] 1100  100    98.33029      7755.357           0   6     20       6
#>  [3,]  700  700    98.33029      7755.357           0   6     20      34
#>  [4,]  500  900    98.33029      7755.357           0   6     20      43
#>  [5,] 1700  900    98.33029      7755.357         100   6     20      49
#>  [6,]  100 1300    98.33029      7755.357           0   6     20      61
#>  [7,]  300 1500    98.33029      7755.357           0   6     20      72
#>  [8,] 1300 1700    98.33029      7755.357           0   6     20      87
#>  [9,] 1700 1700    98.33029      7755.357           0   6     20      89
#> [10,] 1500 1900    98.33029      7755.357           0   6     20      98
#>       Parkfitness
#>  [1,]    7755.357
#>  [2,]    7755.357
#>  [3,]    7755.357
#>  [4,]    7755.357
#>  [5,]    7755.357
#>  [6,]    7755.357
#>  [7,]    7755.357
#>  [8,]    7755.357
#>  [9,]    7755.357
#> [10,]    7755.357
#> 

# }