Get the Grid-IDs from binary matrix

Get the grid IDs from the trimmed binary matrix, where the binary code indicates which grid cells are used in the current wind farm constellation.

get_grids(trimtonOut, Grid)

Arguments

trimtonOut
Input matrix with binary values.

Grid
Grid of the considered area

Value

Returns a list of all individuals with X and Y coordinates and the grid cell ID.

See also

Examples

## Create a random rectangular shapefile
library(sp)
Polygon1 <- Polygon(rbind(c(0, 0), c(0, 2000), c(2000, 2000), c(2000, 0)))
Polygon1 <- Polygons(list(Polygon1),1);
Polygon1 <- SpatialPolygons(list(Polygon1))
Projection <- "+proj=laea +lat_0=52 +lon_0=10 +x_0=4321000 +y_0=3210000
+ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs"
proj4string(Polygon1) <- CRS(Projection)

## Calculate a Grid and an indexed data.frame with coordinates and grid cell Ids.
Grid1 <- grid_area(shape = Polygon1,resol = 200,prop = 1);
Grid <- Grid1[[1]]
AmountGrids <- nrow(Grid)

startsel <- init_population(Grid,10,20);
wind <- data.frame(ws = 12, wd = 0)
wind <- list(wind, probab = 100)
fit <- fitness(selection = startsel,referenceHeight = 100, RotorHeight=100,
               SurfaceRoughness=0.3,Polygon = Polygon1, resol1 = 200,rot=20,
               dirspeed = wind, srtm_crop="",topograp=FALSE,cclRaster="")
allparks <- do.call("rbind",fit);

## SELECTION
## print the amount of Individuals selected.
## Check if the amount of Turbines is as requested.
selec6best <- selection(fit, Grid,2, TRUE, 6, "VAR");

## CROSSOVER
## u determines the amount of crossover points,
## crossPart determines the method used (Equal/Random),
## uplimit is the maximum allowed permutations
crossOut <- crossover(selec6best, 2, uplimit = 300, crossPart="RAN");

## MUTATION
## Variable Mutation Rate is activated if more than 2 individuals represent the
## current best solution.
mut <- mutation(a = crossOut, p = 0.3);

## TRIMTON
## After Crossover and Mutation, the amount of turbines in a windpark change
## and have to be corrected to the required amount of turbines.
mut1 <- trimton(mut = mut, nturb = 10, allparks = allparks,
               nGrids = AmountGrids, trimForce=FALSE)

## Get the new Grid-Ids and run a new fitness run.
getRectV <- get_grids(mut1, Grid)
fit <- fitness(selection = getRectV,referenceHeight = 100, RotorHeight=100,
               SurfaceRoughness=0.3,Polygon = Polygon1, resol1 = 200,rot=20,
               dirspeed = wind, srtm_crop="",topograp=FALSE,cclRaster="")
head(fit)

#> $`1,7,29,31,44,50,53,66,99,100`
#>        X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#> 1    100 1900    96.91844      7644.003           0   1     20       1
#> 7   1300 1900    96.91844      7644.003           0   1     20       7
#> 29  1700 1500    96.91844      7644.003           0   1     20      29
#> 31   100 1300    96.91844      7644.003         100   1     20      31
#> 44   700 1100    96.91844      7644.003           0   1     20      44
#> 50  1900 1100    96.91844      7644.003           0   1     20      50
#> 53   500  900    96.91844      7644.003           0   1     20      53
#> 66  1100  700    96.91844      7644.003           0   1     20      66
#> 99  1700  100    96.91844      7644.003         100   1     20      99
#> 100 1900  100    96.91844      7644.003         100   1     20     100
#>     Parkfitness
#> 1      7644.003
#> 7      7644.003
#> 29     7644.003
#> 31     7644.003
#> 44     7644.003
#> 50     7644.003
#> 53     7644.003
#> 66     7644.003
#> 99     7644.003
#> 100    7644.003
#> 
#> $`1,6,35,36,51,54,61,71,88,90`
#>       X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#> 1   100 1900    86.67868      6836.388           0   2     20       1
#> 6  1100 1900    86.67868      6836.388           0   2     20       6
#> 35  900 1300    86.67868      6836.388           0   2     20      35
#> 36 1100 1300    86.67868      6836.388         100   2     20      36
#> 51  100  900    86.67868      6836.388         100   2     20      51
#> 54  700  900    86.67868      6836.388           0   2     20      54
#> 61  100  700    86.67868      6836.388         200   2     20      61
#> 71  100  500    86.67868      6836.388         300   2     20      71
#> 88 1500  300    86.67868      6836.388           0   2     20      88
#> 90 1900  300    86.67868      6836.388           0   2     20      90
#>    Parkfitness
#> 1     6836.388
#> 6     6836.388
#> 35    6836.388
#> 36    6836.388
#> 51    6836.388
#> 54    6836.388
#> 61    6836.388
#> 71    6836.388
#> 88    6836.388
#> 90    6836.388
#> 
#> $`4,6,7,11,19,20,64,69,92,99`
#>       X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#> 4   700 1900    96.74029      7629.952           0   3     20       4
#> 6  1100 1900    96.74029      7629.952           0   3     20       6
#> 7  1300 1900    96.74029      7629.952           0   3     20       7
#> 11  100 1700    96.74029      7629.952           0   3     20      11
#> 19 1700 1700    96.74029      7629.952           0   3     20      19
#> 20 1900 1700    96.74029      7629.952           0   3     20      20
#> 64  700  700    96.74029      7629.952         100   3     20      64
#> 69 1700  700    96.74029      7629.952         100   3     20      69
#> 92  300  100    96.74029      7629.952           0   3     20      92
#> 99 1700  100    96.74029      7629.952         200   3     20      99
#>    Parkfitness
#> 4     7629.952
#> 6     7629.952
#> 7     7629.952
#> 11    7629.952
#> 19    7629.952
#> 20    7629.952
#> 64    7629.952
#> 69    7629.952
#> 92    7629.952
#> 99    7629.952
#> 
#> $`7,11,13,16,22,28,47,54,74,97`
#>       X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#> 7  1300 1900     95.1034       7500.85           0   4     20       7
#> 11  100 1700     95.1034       7500.85           0   4     20      11
#> 13  500 1700     95.1034       7500.85           0   4     20      13
#> 16 1100 1700     95.1034       7500.85           0   4     20      16
#> 22  300 1500     95.1034       7500.85           0   4     20      22
#> 28 1500 1500     95.1034       7500.85           0   4     20      28
#> 47 1300 1100     95.1034       7500.85         100   4     20      47
#> 54  700  900     95.1034       7500.85           0   4     20      54
#> 74  700  500     95.1034       7500.85         100   4     20      74
#> 97 1300  100     95.1034       7500.85         200   4     20      97
#>    Parkfitness
#> 7      7500.85
#> 11     7500.85
#> 13     7500.85
#> 16     7500.85
#> 22     7500.85
#> 28     7500.85
#> 47     7500.85
#> 54     7500.85
#> 74     7500.85
#> 97     7500.85
#> 
#> $`14,20,26,29,32,36,47,51,61,89`
#>       X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#> 14  700 1700    89.09909      7027.287           0   5     20      14
#> 20 1900 1700    89.09909      7027.287           0   5     20      20
#> 26 1100 1500    89.09909      7027.287           0   5     20      26
#> 29 1700 1500    89.09909      7027.287           0   5     20      29
#> 32  300 1300    89.09909      7027.287           0   5     20      32
#> 36 1100 1300    89.09909      7027.287         100   5     20      36
#> 47 1300 1100    89.09909      7027.287           0   5     20      47
#> 51  100  900    89.09909      7027.287           0   5     20      51
#> 61  100  700    89.09909      7027.287         100   5     20      61
#> 89 1700  300    89.09909      7027.287         100   5     20      89
#>    Parkfitness
#> 14    7027.287
#> 20    7027.287
#> 26    7027.287
#> 29    7027.287
#> 32    7027.287
#> 36    7027.287
#> 47    7027.287
#> 51    7027.287
#> 61    7027.287
#> 89    7027.287
#> 
#> $`3,13,26,32,33,45,54,61,64,77`
#>       X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#> 3   500 1900    86.51307      6823.327           0   6     20       3
#> 13  500 1700    86.51307      6823.327         100   6     20      13
#> 26 1100 1500    86.51307      6823.327           0   6     20      26
#> 32  300 1300    86.51307      6823.327           0   6     20      32
#> 33  500 1300    86.51307      6823.327         200   6     20      33
#> 45  900 1100    86.51307      6823.327           0   6     20      45
#> 54  700  900    86.51307      6823.327           0   6     20      54
#> 61  100  700    86.51307      6823.327           0   6     20      61
#> 64  700  700    86.51307      6823.327         100   6     20      64
#> 77 1300  500    86.51307      6823.327           0   6     20      77
#>    Parkfitness
#> 3     6823.327
#> 13    6823.327
#> 26    6823.327
#> 32    6823.327
#> 33    6823.327
#> 45    6823.327
#> 54    6823.327
#> 61    6823.327
#> 64    6823.327
#> 77    6823.327
#>