Get the Grid-IDs from binary matrix

Get the grid IDs from the trimmed binary matrix, where the binary code indicates which grid cells are used in the current wind farm constellation.

get_grids(trimtonOut, Grid)

Arguments

trimtonOut
Input matrix with binary values.

Grid
Grid of the considered area

Value

Returns a list of all individuals with X and Y coordinates and the grid cell ID.

See also

Examples

## Create a random rectangular shapefile
library(sp)
Polygon1 <- Polygon(rbind(c(0, 0), c(0, 2000), c(2000, 2000), c(2000, 0)))
Polygon1 <- Polygons(list(Polygon1),1);
Polygon1 <- SpatialPolygons(list(Polygon1))
Projection <- "+proj=laea +lat_0=52 +lon_0=10 +x_0=4321000 +y_0=3210000
+ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs"
proj4string(Polygon1) <- CRS(Projection)

## Calculate a Grid and an indexed data.frame with coordinates and grid cell Ids.
Grid1 <- grid_area(shape = Polygon1,resol = 200,prop = 1);
Grid <- Grid1[[1]]
AmountGrids <- nrow(Grid)

startsel <- init_population(Grid,10,20);
wind <- data.frame(ws = 12, wd = 0)
wind <- list(wind, probab = 100)
fit <- fitness(selection = startsel,referenceHeight = 100, RotorHeight=100,
               SurfaceRoughness=0.3,Polygon = Polygon1, resol1 = 200,rot=20,
               dirspeed = wind, srtm_crop="",topograp=FALSE,cclRaster="")
allparks <- do.call("rbind",fit);

## SELECTION
## print the amount of Individuals selected.
## Check if the amount of Turbines is as requested.
selec6best <- selection(fit, Grid,2, TRUE, 6, "VAR");

## CROSSOVER
## u determines the amount of crossover points,
## crossPart determines the method used (Equal/Random),
## uplimit is the maximum allowed permutations
crossOut <- crossover(selec6best, 2, uplimit = 300, crossPart="RAN");

## MUTATION
## Variable Mutation Rate is activated if more than 2 individuals represent the
## current best solution.
mut <- mutation(a = crossOut, p = 0.3);

## TRIMTON
## After Crossover and Mutation, the amount of turbines in a windpark change
## and have to be corrected to the required amount of turbines.
mut1 <- trimton(mut = mut, nturb = 10, allparks = allparks,
               nGrids = AmountGrids, trimForce=FALSE)

## Get the new Grid-Ids and run a new fitness run.
getRectV <- get_grids(mut1, Grid)
fit <- fitness(selection = getRectV,referenceHeight = 100, RotorHeight=100,
               SurfaceRoughness=0.3,Polygon = Polygon1, resol1 = 200,rot=20,
               dirspeed = wind, srtm_crop="",topograp=FALSE,cclRaster="")
head(fit)

#> $`31,32,36,56,59,80,82,85,93,94`
#>       X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#> 31  100 1300     96.3353       7598.01           0   1     20      31
#> 32  300 1300     96.3353       7598.01           0   1     20      32
#> 36 1100 1300     96.3353       7598.01           0   1     20      36
#> 56 1100  900     96.3353       7598.01         100   1     20      56
#> 59 1700  900     96.3353       7598.01           0   1     20      59
#> 80 1900  500     96.3353       7598.01           0   1     20      80
#> 82  300  300     96.3353       7598.01         100   1     20      82
#> 85  900  300     96.3353       7598.01           0   1     20      85
#> 93  500  100     96.3353       7598.01           0   1     20      93
#> 94  700  100     96.3353       7598.01           0   1     20      94
#>    Parkfitness
#> 31     7598.01
#> 32     7598.01
#> 36     7598.01
#> 56     7598.01
#> 59     7598.01
#> 80     7598.01
#> 82     7598.01
#> 85     7598.01
#> 93     7598.01
#> 94     7598.01
#> 
#> $`28,30,43,45,56,62,66,67,80,98`
#>       X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#> 28 1500 1500    93.52897      7376.674           0   2     20      28
#> 30 1900 1500    93.52897      7376.674           0   2     20      30
#> 43  500 1100    93.52897      7376.674           0   2     20      43
#> 45  900 1100    93.52897      7376.674           0   2     20      45
#> 56 1100  900    93.52897      7376.674           0   2     20      56
#> 62  300  700    93.52897      7376.674           0   2     20      62
#> 66 1100  700    93.52897      7376.674         100   2     20      66
#> 67 1300  700    93.52897      7376.674           0   2     20      67
#> 80 1900  500    93.52897      7376.674         100   2     20      80
#> 98 1500  100    93.52897      7376.674         100   2     20      98
#>    Parkfitness
#> 28    7376.674
#> 30    7376.674
#> 43    7376.674
#> 45    7376.674
#> 56    7376.674
#> 62    7376.674
#> 66    7376.674
#> 67    7376.674
#> 80    7376.674
#> 98    7376.674
#> 
#> $`21,38,52,53,63,66,68,70,76,84`
#>       X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#> 21  100 1500    87.98952      6939.775           0   3     20      21
#> 38 1500 1300    87.98952      6939.775           0   3     20      38
#> 52  300  900    87.98952      6939.775           0   3     20      52
#> 53  500  900    87.98952      6939.775           0   3     20      53
#> 63  500  700    87.98952      6939.775         100   3     20      63
#> 66 1100  700    87.98952      6939.775           0   3     20      66
#> 68 1500  700    87.98952      6939.775         100   3     20      68
#> 70 1900  700    87.98952      6939.775           0   3     20      70
#> 76 1100  500    87.98952      6939.775         100   3     20      76
#> 84  700  300    87.98952      6939.775           0   3     20      84
#>    Parkfitness
#> 21    6939.775
#> 38    6939.775
#> 52    6939.775
#> 53    6939.775
#> 63    6939.775
#> 66    6939.775
#> 68    6939.775
#> 70    6939.775
#> 76    6939.775
#> 84    6939.775
#> 
#> $`2,19,20,23,24,28,30,75,87,93`
#>       X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#> 2   300 1900    94.37308       7443.25           0   4     20       2
#> 19 1700 1700    94.37308       7443.25           0   4     20      19
#> 20 1900 1700    94.37308       7443.25           0   4     20      20
#> 23  500 1500    94.37308       7443.25           0   4     20      23
#> 24  700 1500    94.37308       7443.25           0   4     20      24
#> 28 1500 1500    94.37308       7443.25           0   4     20      28
#> 30 1900 1500    94.37308       7443.25         100   4     20      30
#> 75  900  500    94.37308       7443.25           0   4     20      75
#> 87 1300  300    94.37308       7443.25           0   4     20      87
#> 93  500  100    94.37308       7443.25         100   4     20      93
#>    Parkfitness
#> 2      7443.25
#> 19     7443.25
#> 20     7443.25
#> 23     7443.25
#> 24     7443.25
#> 28     7443.25
#> 30     7443.25
#> 75     7443.25
#> 87     7443.25
#> 93     7443.25
#> 
#> $`1,13,15,38,42,48,57,64,91,99`
#>       X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#> 1   100 1900    94.54377      7456.712           0   5     20       1
#> 13  500 1700    94.54377      7456.712           0   5     20      13
#> 15  900 1700    94.54377      7456.712           0   5     20      15
#> 38 1500 1300    94.54377      7456.712           0   5     20      38
#> 42  300 1100    94.54377      7456.712           0   5     20      42
#> 48 1500 1100    94.54377      7456.712         100   5     20      48
#> 57 1300  900    94.54377      7456.712           0   5     20      57
#> 64  700  700    94.54377      7456.712           0   5     20      64
#> 91  100  100    94.54377      7456.712         100   5     20      91
#> 99 1700  100    94.54377      7456.712           0   5     20      99
#>    Parkfitness
#> 1     7456.712
#> 13    7456.712
#> 15    7456.712
#> 38    7456.712
#> 42    7456.712
#> 48    7456.712
#> 57    7456.712
#> 64    7456.712
#> 91    7456.712
#> 99    7456.712
#> 
#> $`25,38,43,47,55,62,66,80,85,100`
#>        X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#> 25   900 1500    93.59535      7381.909           0   6     20      25
#> 38  1500 1300    93.59535      7381.909           0   6     20      38
#> 43   500 1100    93.59535      7381.909           0   6     20      43
#> 47  1300 1100    93.59535      7381.909           0   6     20      47
#> 55   900  900    93.59535      7381.909         100   6     20      55
#> 62   300  700    93.59535      7381.909           0   6     20      62
#> 66  1100  700    93.59535      7381.909           0   6     20      66
#> 80  1900  500    93.59535      7381.909           0   6     20      80
#> 85   900  300    93.59535      7381.909         200   6     20      85
#> 100 1900  100    93.59535      7381.909         100   6     20     100
#>     Parkfitness
#> 25     7381.909
#> 38     7381.909
#> 43     7381.909
#> 47     7381.909
#> 55     7381.909
#> 62     7381.909
#> 66     7381.909
#> 80     7381.909
#> 85     7381.909
#> 100    7381.909
#>