Get the Grid-IDs from binary matrix

Retrieve the grid ID's from the binary matrix, where the binary code indicates which grid cells are used in the current wind farm constellation.

get_grids(trimtonOut, Grid)

Arguments

trimtonOut

Input matrix with binary values

Grid

Grid of the considered area

Value

Returns a list of all individuals with X and Y coordinates and the grid cell ID.

See also

Examples

# \donttest{
## Create a random rectangular shapefile
library(sf)
Polygon1 <- sf::st_as_sf(sf::st_sfc(
 sf::st_polygon(list(cbind(
   c(0, 0, 2000, 2000, 0),
   c(0, 2000, 2000, 0, 0)))),
crs = 3035
))

## Calculate a Grid and an indexed data.frame with coordinates and grid cell Ids.
Grid1 <- grid_area(shape = Polygon1, size = 200, prop = 1)
Grid <- Grid1[[1]]
AmountGrids <- nrow(Grid)

startsel <- init_population(Grid,10,20)
wind <- data.frame(ws = 12, wd = 0)
wind <- list(wind, probab = 100)
fit <- fitness(selection = startsel,referenceHeight = 100, RotorHeight=100,
               SurfaceRoughness=0.3,Polygon = Polygon1, resol1 = 200,rot=20,
               dirspeed = wind, srtm_crop="",topograp=FALSE,cclRaster="")
allparks <- do.call("rbind",fit)

## SELECTION
## print the amount of Individuals selected.
## Check if the amount of Turbines is as requested.
selec6best <- selection(fit, Grid,2, TRUE, 6, "VAR")

## CROSSOVER
## u determines the amount of crossover points,
## crossPart determines the method used (Equal/Random),
## uplimit is the maximum allowed permutations
crossOut <- crossover(selec6best, 2, uplimit = 300, crossPart="RAN")

## MUTATION
## Variable Mutation Rate is activated if more than 2 individuals represent the
## current best solution.
mut <- mutation(a = crossOut, p = 0.3)

## TRIMTON
## After Crossover and Mutation, the amount of turbines in a windpark change
## and have to be corrected to the required amount of turbines.
mut1 <- trimton(mut = mut, nturb = 10, allparks = allparks,
               nGrids = AmountGrids, trimForce=FALSE)

## Get the new Grid-Ids and run a new fitness run.
getRectV <- get_grids(mut1, Grid)
fit <- fitness(selection = getRectV,referenceHeight = 100, RotorHeight=100,
               SurfaceRoughness=0.3,Polygon = Polygon1, resol1 = 200,rot=20,
               dirspeed = wind, srtm_crop="",topograp=FALSE,cclRaster="")
head(fit)
#> $`3,7,9,25,28,40,49,54,70,79`
#>          X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#>  [1,]  500  100     95.2456      7512.065           0   1     20       3
#>  [2,] 1300  100     95.2456      7512.065           0   1     20       7
#>  [3,] 1700  100     95.2456      7512.065         200   1     20       9
#>  [4,]  900  500     95.2456      7512.065           0   1     20      25
#>  [5,] 1500  500     95.2456      7512.065           0   1     20      28
#>  [6,] 1900  700     95.2456      7512.065         100   1     20      40
#>  [7,] 1700  900     95.2456      7512.065         100   1     20      49
#>  [8,]  700 1100     95.2456      7512.065           0   1     20      54
#>  [9,] 1900 1300     95.2456      7512.065           0   1     20      70
#> [10,] 1700 1500     95.2456      7512.065           0   1     20      79
#>       Parkfitness
#>  [1,]    7512.065
#>  [2,]    7512.065
#>  [3,]    7512.065
#>  [4,]    7512.065
#>  [5,]    7512.065
#>  [6,]    7512.065
#>  [7,]    7512.065
#>  [8,]    7512.065
#>  [9,]    7512.065
#> [10,]    7512.065
#> 
#> $`4,21,23,28,40,50,59,73,79,87`
#>          X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#>  [1,]  700  100    91.20199      7193.144           0   2     20       4
#>  [2,]  100  500    91.20199      7193.144           0   2     20      21
#>  [3,]  500  500    91.20199      7193.144         100   2     20      23
#>  [4,] 1500  500    91.20199      7193.144           0   2     20      28
#>  [5,] 1900  700    91.20199      7193.144         100   2     20      40
#>  [6,] 1900  900    91.20199      7193.144           0   2     20      50
#>  [7,] 1700 1100    91.20199      7193.144         100   2     20      59
#>  [8,]  500 1500    91.20199      7193.144           0   2     20      73
#>  [9,] 1700 1500    91.20199      7193.144           0   2     20      79
#> [10,] 1300 1700    91.20199      7193.144           0   2     20      87
#>       Parkfitness
#>  [1,]    7193.144
#>  [2,]    7193.144
#>  [3,]    7193.144
#>  [4,]    7193.144
#>  [5,]    7193.144
#>  [6,]    7193.144
#>  [7,]    7193.144
#>  [8,]    7193.144
#>  [9,]    7193.144
#> [10,]    7193.144
#> 
#> $`16,21,27,28,46,54,58,77,88,96`
#>          X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#>  [1,] 1100  300    92.94641      7330.727         200   3     20      16
#>  [2,]  100  500    92.94641      7330.727           0   3     20      21
#>  [3,] 1300  500    92.94641      7330.727         100   3     20      27
#>  [4,] 1500  500    92.94641      7330.727         200   3     20      28
#>  [5,] 1100  900    92.94641      7330.727         100   3     20      46
#>  [6,]  700 1100    92.94641      7330.727           0   3     20      54
#>  [7,] 1500 1100    92.94641      7330.727         100   3     20      58
#>  [8,] 1300 1500    92.94641      7330.727           0   3     20      77
#>  [9,] 1500 1700    92.94641      7330.727           0   3     20      88
#> [10,] 1100 1900    92.94641      7330.727           0   3     20      96
#>       Parkfitness
#>  [1,]    7330.727
#>  [2,]    7330.727
#>  [3,]    7330.727
#>  [4,]    7330.727
#>  [5,]    7330.727
#>  [6,]    7330.727
#>  [7,]    7330.727
#>  [8,]    7330.727
#>  [9,]    7330.727
#> [10,]    7330.727
#> 
#> $`29,30,35,42,46,48,50,51,56,84`
#>          X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#>  [1,] 1700  500    92.04609      7259.719           0   4     20      29
#>  [2,] 1900  500    92.04609      7259.719         100   4     20      30
#>  [3,]  900  700    92.04609      7259.719           0   4     20      35
#>  [4,]  300  900    92.04609      7259.719           0   4     20      42
#>  [5,] 1100  900    92.04609      7259.719         100   4     20      46
#>  [6,] 1500  900    92.04609      7259.719           0   4     20      48
#>  [7,] 1900  900    92.04609      7259.719           0   4     20      50
#>  [8,]  100 1100    92.04609      7259.719           0   4     20      51
#>  [9,] 1100 1100    92.04609      7259.719           0   4     20      56
#> [10,]  700 1700    92.04609      7259.719           0   4     20      84
#>       Parkfitness
#>  [1,]    7259.719
#>  [2,]    7259.719
#>  [3,]    7259.719
#>  [4,]    7259.719
#>  [5,]    7259.719
#>  [6,]    7259.719
#>  [7,]    7259.719
#>  [8,]    7259.719
#>  [9,]    7259.719
#> [10,]    7259.719
#> 
#> $`9,16,46,60,62,70,77,80,89,96`
#>          X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#>  [1,] 1700  100    86.27683      6804.695         100   5     20       9
#>  [2,] 1100  300    86.27683      6804.695         200   5     20      16
#>  [3,] 1100  900    86.27683      6804.695         100   5     20      46
#>  [4,] 1900 1100    86.27683      6804.695         200   5     20      60
#>  [5,]  300 1300    86.27683      6804.695           0   5     20      62
#>  [6,] 1900 1300    86.27683      6804.695         100   5     20      70
#>  [7,] 1300 1500    86.27683      6804.695           0   5     20      77
#>  [8,] 1900 1500    86.27683      6804.695           0   5     20      80
#>  [9,] 1700 1700    86.27683      6804.695           0   5     20      89
#> [10,] 1100 1900    86.27683      6804.695           0   5     20      96
#>       Parkfitness
#>  [1,]    6804.695
#>  [2,]    6804.695
#>  [3,]    6804.695
#>  [4,]    6804.695
#>  [5,]    6804.695
#>  [6,]    6804.695
#>  [7,]    6804.695
#>  [8,]    6804.695
#>  [9,]    6804.695
#> [10,]    6804.695
#> 
#> $`20,22,25,35,45,50,67,78,83,92`
#>          X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#>  [1,] 1900  300    87.05965      6866.436         100   6     20      20
#>  [2,]  300  500    87.05965      6866.436         100   6     20      22
#>  [3,]  900  500    87.05965      6866.436         200   6     20      25
#>  [4,]  900  700    87.05965      6866.436         100   6     20      35
#>  [5,]  900  900    87.05965      6866.436           0   6     20      45
#>  [6,] 1900  900    87.05965      6866.436           0   6     20      50
#>  [7,] 1300 1300    87.05965      6866.436           0   6     20      67
#>  [8,] 1500 1500    87.05965      6866.436           0   6     20      78
#>  [9,]  500 1700    87.05965      6866.436           0   6     20      83
#> [10,]  300 1900    87.05965      6866.436           0   6     20      92
#>       Parkfitness
#>  [1,]    6866.436
#>  [2,]    6866.436
#>  [3,]    6866.436
#>  [4,]    6866.436
#>  [5,]    6866.436
#>  [6,]    6866.436
#>  [7,]    6866.436
#>  [8,]    6866.436
#>  [9,]    6866.436
#> [10,]    6866.436
#> 

# }