Get the Grid-IDs from binary matrix

Get the grid IDs from the trimmed binary matrix, where the binary code indicates which grid cells are used in the current wind farm constellation.

get_grids(trimtonOut, Grid)

Arguments

trimtonOut
Input matrix with binary values.

Grid
Grid of the considered area

Value

Returns a list of all individuals with X and Y coordinates and the grid cell ID.

See also

Examples

# \donttest{
## Create a random rectangular shapefile
library(sp)
Polygon1 <- Polygon(rbind(c(0, 0), c(0, 2000), c(2000, 2000), c(2000, 0)))
Polygon1 <- Polygons(list(Polygon1),1);
Polygon1 <- SpatialPolygons(list(Polygon1))
Projection <- "+proj=laea +lat_0=52 +lon_0=10 +x_0=4321000 +y_0=3210000
+ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs"
proj4string(Polygon1) <- CRS(Projection)

## Calculate a Grid and an indexed data.frame with coordinates and grid cell Ids.
Grid1 <- grid_area(shape = Polygon1,resol = 200,prop = 1);
Grid <- Grid1[[1]]
AmountGrids <- nrow(Grid)

startsel <- init_population(Grid,10,20);
wind <- data.frame(ws = 12, wd = 0)
wind <- list(wind, probab = 100)
fit <- fitness(selection = startsel,referenceHeight = 100, RotorHeight=100,
               SurfaceRoughness=0.3,Polygon = Polygon1, resol1 = 200,rot=20,
               dirspeed = wind, srtm_crop="",topograp=FALSE,cclRaster="")
allparks <- do.call("rbind",fit);

## SELECTION
## print the amount of Individuals selected.
## Check if the amount of Turbines is as requested.
selec6best <- selection(fit, Grid,2, TRUE, 6, "VAR");

## CROSSOVER
## u determines the amount of crossover points,
## crossPart determines the method used (Equal/Random),
## uplimit is the maximum allowed permutations
crossOut <- crossover(selec6best, 2, uplimit = 300, crossPart="RAN");

## MUTATION
## Variable Mutation Rate is activated if more than 2 individuals represent the
## current best solution.
mut <- mutation(a = crossOut, p = 0.3);

## TRIMTON
## After Crossover and Mutation, the amount of turbines in a windpark change
## and have to be corrected to the required amount of turbines.
mut1 <- trimton(mut = mut, nturb = 10, allparks = allparks,
               nGrids = AmountGrids, trimForce=FALSE)

## Get the new Grid-Ids and run a new fitness run.
getRectV <- get_grids(mut1, Grid)
fit <- fitness(selection = getRectV,referenceHeight = 100, RotorHeight=100,
               SurfaceRoughness=0.3,Polygon = Polygon1, resol1 = 200,rot=20,
               dirspeed = wind, srtm_crop="",topograp=FALSE,cclRaster="")
head(fit)

#> $`14,19,21,45,50,51,58,75,98,99`
#>       X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#> 14  700 1700    94.94135      7488.069           0   1     20      14
#> 19 1700 1700    94.94135      7488.069           0   1     20      19
#> 21  100 1500    94.94135      7488.069           0   1     20      21
#> 45  900 1100    94.94135      7488.069           0   1     20      45
#> 50 1900 1100    94.94135      7488.069           0   1     20      50
#> 51  100  900    94.94135      7488.069         100   1     20      51
#> 58 1500  900    94.94135      7488.069           0   1     20      58
#> 75  900  500    94.94135      7488.069         100   1     20      75
#> 98 1500  100    94.94135      7488.069         100   1     20      98
#> 99 1700  100    94.94135      7488.069         100   1     20      99
#>    Parkfitness
#> 14    7488.069
#> 19    7488.069
#> 21    7488.069
#> 45    7488.069
#> 50    7488.069
#> 51    7488.069
#> 58    7488.069
#> 75    7488.069
#> 98    7488.069
#> 99    7488.069
#> 
#> $`5,15,19,44,47,58,67,78,92,96`
#>       X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#> 5   900 1900     89.2255      7037.257           0   2     20       5
#> 15  900 1700     89.2255      7037.257         100   2     20      15
#> 19 1700 1700     89.2255      7037.257           0   2     20      19
#> 44  700 1100     89.2255      7037.257           0   2     20      44
#> 47 1300 1100     89.2255      7037.257           0   2     20      47
#> 58 1500  900     89.2255      7037.257           0   2     20      58
#> 67 1300  700     89.2255      7037.257         100   2     20      67
#> 78 1500  500     89.2255      7037.257         100   2     20      78
#> 92  300  100     89.2255      7037.257           0   2     20      92
#> 96 1100  100     89.2255      7037.257           0   2     20      96
#>    Parkfitness
#> 5     7037.257
#> 15    7037.257
#> 19    7037.257
#> 44    7037.257
#> 47    7037.257
#> 58    7037.257
#> 67    7037.257
#> 78    7037.257
#> 92    7037.257
#> 96    7037.257
#> 
#> $`19,23,25,33,66,67,77,85,86,97`
#>       X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#> 19 1700 1700    83.05819      6550.839           0   3     20      19
#> 23  500 1500    83.05819      6550.839           0   3     20      23
#> 25  900 1500    83.05819      6550.839           0   3     20      25
#> 33  500 1300    83.05819      6550.839         100   3     20      33
#> 66 1100  700    83.05819      6550.839           0   3     20      66
#> 67 1300  700    83.05819      6550.839           0   3     20      67
#> 77 1300  500    83.05819      6550.839         100   3     20      77
#> 85  900  300    83.05819      6550.839         100   3     20      85
#> 86 1100  300    83.05819      6550.839         100   3     20      86
#> 97 1300  100    83.05819      6550.839         200   3     20      97
#>    Parkfitness
#> 19    6550.839
#> 23    6550.839
#> 25    6550.839
#> 33    6550.839
#> 66    6550.839
#> 67    6550.839
#> 77    6550.839
#> 85    6550.839
#> 86    6550.839
#> 97    6550.839
#> 
#> $`5,17,23,24,29,30,32,39,50,67`
#>       X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#> 5   900 1900    91.20199      7193.144           0   4     20       5
#> 17 1300 1700    91.20199      7193.144           0   4     20      17
#> 23  500 1500    91.20199      7193.144           0   4     20      23
#> 24  700 1500    91.20199      7193.144           0   4     20      24
#> 29 1700 1500    91.20199      7193.144           0   4     20      29
#> 30 1900 1500    91.20199      7193.144           0   4     20      30
#> 32  300 1300    91.20199      7193.144           0   4     20      32
#> 39 1700 1300    91.20199      7193.144         100   4     20      39
#> 50 1900 1100    91.20199      7193.144         100   4     20      50
#> 67 1300  700    91.20199      7193.144         100   4     20      67
#>    Parkfitness
#> 5     7193.144
#> 17    7193.144
#> 23    7193.144
#> 24    7193.144
#> 29    7193.144
#> 30    7193.144
#> 32    7193.144
#> 39    7193.144
#> 50    7193.144
#> 67    7193.144
#> 
#> $`2,20,21,27,33,34,43,62,68,80`
#>       X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#> 2   300 1900    93.59812      7382.128           0   5     20       2
#> 20 1900 1700    93.59812      7382.128           0   5     20      20
#> 21  100 1500    93.59812      7382.128           0   5     20      21
#> 27 1300 1500    93.59812      7382.128           0   5     20      27
#> 33  500 1300    93.59812      7382.128           0   5     20      33
#> 34  700 1300    93.59812      7382.128           0   5     20      34
#> 43  500 1100    93.59812      7382.128         100   5     20      43
#> 62  300  700    93.59812      7382.128         100   5     20      62
#> 68 1500  700    93.59812      7382.128           0   5     20      68
#> 80 1900  500    93.59812      7382.128         100   5     20      80
#>    Parkfitness
#> 2     7382.128
#> 20    7382.128
#> 21    7382.128
#> 27    7382.128
#> 33    7382.128
#> 34    7382.128
#> 43    7382.128
#> 62    7382.128
#> 68    7382.128
#> 80    7382.128
#> 
#> $`1,2,5,6,13,27,32,33,65,77`
#>       X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#> 1   100 1900    93.95716      7410.446           0   6     20       1
#> 2   300 1900    93.95716      7410.446           0   6     20       2
#> 5   900 1900    93.95716      7410.446           0   6     20       5
#> 6  1100 1900    93.95716      7410.446           0   6     20       6
#> 13  500 1700    93.95716      7410.446           0   6     20      13
#> 27 1300 1500    93.95716      7410.446           0   6     20      27
#> 32  300 1300    93.95716      7410.446         100   6     20      32
#> 33  500 1300    93.95716      7410.446         100   6     20      33
#> 65  900  700    93.95716      7410.446         100   6     20      65
#> 77 1300  500    93.95716      7410.446         100   6     20      77
#>    Parkfitness
#> 1     7410.446
#> 2     7410.446
#> 5     7410.446
#> 6     7410.446
#> 13    7410.446
#> 27    7410.446
#> 32    7410.446
#> 33    7410.446
#> 65    7410.446
#> 77    7410.446
#> 

# }