Get the Grid-IDs from binary matrix

Get the grid IDs from the trimmed binary matrix, where the binary code indicates which grid cells are used in the current wind farm constellation.

get_grids(trimtonOut, Grid)

Arguments

trimtonOut
Input matrix with binary values.

Grid
Grid of the considered area

Value

Returns a list of all individuals with X and Y coordinates and the grid cell ID.

See also

Examples

## Create a random rectangular shapefile
library(sp)
Polygon1 <- Polygon(rbind(c(0, 0), c(0, 2000), c(2000, 2000), c(2000, 0)))
Polygon1 <- Polygons(list(Polygon1),1);
Polygon1 <- SpatialPolygons(list(Polygon1))
Projection <- "+proj=laea +lat_0=52 +lon_0=10 +x_0=4321000 +y_0=3210000
+ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs"
proj4string(Polygon1) <- CRS(Projection)

## Calculate a Grid and an indexed data.frame with coordinates and grid cell Ids.
Grid1 <- grid_area(shape = Polygon1,resol = 200,prop = 1);
Grid <- Grid1[[1]]
AmountGrids <- nrow(Grid)

startsel <- init_population(Grid,10,20);
wind <- data.frame(ws = 12, wd = 0)
wind <- list(wind, probab = 100)
fit <- fitness(selection = startsel,referenceHeight = 100, RotorHeight=100,
               SurfaceRoughness=0.3,Polygon = Polygon1, resol1 = 200,rot=20,
               dirspeed = wind, srtm_crop="",topograp=FALSE,cclRaster="")
allparks <- do.call("rbind",fit);

## SELECTION
## print the amount of Individuals selected.
## Check if the amount of Turbines is as requested.
selec6best <- selection(fit, Grid,2, TRUE, 6, "VAR");

## CROSSOVER
## u determines the amount of crossover points,
## crossPart determines the method used (Equal/Random),
## uplimit is the maximum allowed permutations
crossOut <- crossover(selec6best, 2, uplimit = 300, crossPart="RAN");

## MUTATION
## Variable Mutation Rate is activated if more than 2 individuals represent the
## current best solution.
mut <- mutation(a = crossOut, p = 0.3);

## TRIMTON
## After Crossover and Mutation, the amount of turbines in a windpark change
## and have to be corrected to the required amount of turbines.
mut1 <- trimton(mut = mut, nturb = 10, allparks = allparks,
               nGrids = AmountGrids, trimForce=FALSE)

## Get the new Grid-Ids and run a new fitness run.
getRectV <- get_grids(mut1, Grid)
fit <- fitness(selection = getRectV,referenceHeight = 100, RotorHeight=100,
               SurfaceRoughness=0.3,Polygon = Polygon1, resol1 = 200,rot=20,
               dirspeed = wind, srtm_crop="",topograp=FALSE,cclRaster="")
head(fit)

#> $`1,15,19,30,36,61,63,71,81,92`
#>       X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#> 1   100 1900    88.64416      6991.407           0   1     20       1
#> 15  900 1700    88.64416      6991.407           0   1     20      15
#> 19 1700 1700    88.64416      6991.407           0   1     20      19
#> 30 1900 1500    88.64416      6991.407           0   1     20      30
#> 36 1100 1300    88.64416      6991.407           0   1     20      36
#> 61  100  700    88.64416      6991.407         100   1     20      61
#> 63  500  700    88.64416      6991.407           0   1     20      63
#> 71  100  500    88.64416      6991.407         200   1     20      71
#> 81  100  300    88.64416      6991.407         300   1     20      81
#> 92  300  100    88.64416      6991.407           0   1     20      92
#>    Parkfitness
#> 1     6991.407
#> 15    6991.407
#> 19    6991.407
#> 30    6991.407
#> 36    6991.407
#> 61    6991.407
#> 63    6991.407
#> 71    6991.407
#> 81    6991.407
#> 92    6991.407
#> 
#> $`10,17,23,26,55,56,62,63,89,97`
#>       X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#> 10 1900 1900     96.6852      7625.608           0   2     20      10
#> 17 1300 1700     96.6852      7625.608           0   2     20      17
#> 23  500 1500     96.6852      7625.608           0   2     20      23
#> 26 1100 1500     96.6852      7625.608           0   2     20      26
#> 55  900  900     96.6852      7625.608           0   2     20      55
#> 56 1100  900     96.6852      7625.608         100   2     20      56
#> 62  300  700     96.6852      7625.608           0   2     20      62
#> 63  500  700     96.6852      7625.608         100   2     20      63
#> 89 1700  300     96.6852      7625.608           0   2     20      89
#> 97 1300  100     96.6852      7625.608         100   2     20      97
#>    Parkfitness
#> 10    7625.608
#> 17    7625.608
#> 23    7625.608
#> 26    7625.608
#> 55    7625.608
#> 56    7625.608
#> 62    7625.608
#> 63    7625.608
#> 89    7625.608
#> 97    7625.608
#> 
#> $`4,16,25,32,44,53,55,67,84,89`
#>       X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#> 4   700 1900    95.84549      7559.379           0   3     20       4
#> 16 1100 1700    95.84549      7559.379           0   3     20      16
#> 25  900 1500    95.84549      7559.379           0   3     20      25
#> 32  300 1300    95.84549      7559.379           0   3     20      32
#> 44  700 1100    95.84549      7559.379         100   3     20      44
#> 53  500  900    95.84549      7559.379           0   3     20      53
#> 55  900  900    95.84549      7559.379         100   3     20      55
#> 67 1300  700    95.84549      7559.379           0   3     20      67
#> 84  700  300    95.84549      7559.379         200   3     20      84
#> 89 1700  300    95.84549      7559.379           0   3     20      89
#>    Parkfitness
#> 4     7559.379
#> 16    7559.379
#> 25    7559.379
#> 32    7559.379
#> 44    7559.379
#> 53    7559.379
#> 55    7559.379
#> 67    7559.379
#> 84    7559.379
#> 89    7559.379
#> 
#> $`8,16,22,27,29,35,77,85,96,99`
#>       X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#> 8  1500 1900    97.42334      7683.824           0   4     20       8
#> 16 1100 1700    97.42334      7683.824           0   4     20      16
#> 22  300 1500    97.42334      7683.824           0   4     20      22
#> 27 1300 1500    97.42334      7683.824           0   4     20      27
#> 29 1700 1500    97.42334      7683.824           0   4     20      29
#> 35  900 1300    97.42334      7683.824           0   4     20      35
#> 77 1300  500    97.42334      7683.824         100   4     20      77
#> 85  900  300    97.42334      7683.824         100   4     20      85
#> 96 1100  100    97.42334      7683.824         100   4     20      96
#> 99 1700  100    97.42334      7683.824         100   4     20      99
#>    Parkfitness
#> 8     7683.824
#> 16    7683.824
#> 22    7683.824
#> 27    7683.824
#> 29    7683.824
#> 35    7683.824
#> 77    7683.824
#> 85    7683.824
#> 96    7683.824
#> 99    7683.824
#> 
#> $`28,31,37,44,56,60,67,72,75,93`
#>       X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#> 28 1500 1500    98.25615      7749.509           0   5     20      28
#> 31  100 1300    98.25615      7749.509           0   5     20      31
#> 37 1300 1300    98.25615      7749.509           0   5     20      37
#> 44  700 1100    98.25615      7749.509           0   5     20      44
#> 56 1100  900    98.25615      7749.509           0   5     20      56
#> 60 1900  900    98.25615      7749.509           0   5     20      60
#> 67 1300  700    98.25615      7749.509         100   5     20      67
#> 72  300  500    98.25615      7749.509           0   5     20      72
#> 75  900  500    98.25615      7749.509           0   5     20      75
#> 93  500  100    98.25615      7749.509           0   5     20      93
#>    Parkfitness
#> 28    7749.509
#> 31    7749.509
#> 37    7749.509
#> 44    7749.509
#> 56    7749.509
#> 60    7749.509
#> 67    7749.509
#> 72    7749.509
#> 75    7749.509
#> 93    7749.509
#> 
#> $`1,16,27,34,36,53,56,82,92,96`
#>       X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#> 1   100 1900    87.66734      6914.364           0   6     20       1
#> 16 1100 1700    87.66734      6914.364           0   6     20      16
#> 27 1300 1500    87.66734      6914.364           0   6     20      27
#> 34  700 1300    87.66734      6914.364           0   6     20      34
#> 36 1100 1300    87.66734      6914.364         100   6     20      36
#> 53  500  900    87.66734      6914.364           0   6     20      53
#> 56 1100  900    87.66734      6914.364         200   6     20      56
#> 82  300  300    87.66734      6914.364           0   6     20      82
#> 92  300  100    87.66734      6914.364         100   6     20      92
#> 96 1100  100    87.66734      6914.364         300   6     20      96
#>    Parkfitness
#> 1     6914.364
#> 16    6914.364
#> 27    6914.364
#> 34    6914.364
#> 36    6914.364
#> 53    6914.364
#> 56    6914.364
#> 82    6914.364
#> 92    6914.364
#> 96    6914.364
#>