Get the Grid-IDs from binary matrix

Get the grid IDs from the trimmed binary matrix, where the binary code indicates which grid cells are used in the current wind farm constellation.

get_grids(trimtonOut, Grid)

Arguments

trimtonOut
Input matrix with binary values.

Grid
Grid of the considered area

Value

Returns a list of all individuals with X and Y coordinates and the grid cell ID.

See also

Examples

# \donttest{
## Create a random rectangular shapefile
library(sf)
Polygon1 <- sf::st_as_sf(sf::st_sfc(
 sf::st_polygon(list(cbind(
   c(0, 0, 2000, 2000, 0),
   c(0, 2000, 2000, 0, 0)))),
crs = 3035
))

## Calculate a Grid and an indexed data.frame with coordinates and grid cell Ids.
Grid1 <- grid_area(shape = Polygon1, size = 200, prop = 1)
Grid <- Grid1[[1]]
AmountGrids <- nrow(Grid)

startsel <- init_population(Grid,10,20)
wind <- data.frame(ws = 12, wd = 0)
wind <- list(wind, probab = 100)
fit <- fitness(selection = startsel,referenceHeight = 100, RotorHeight=100,
               SurfaceRoughness=0.3,Polygon = Polygon1, resol1 = 200,rot=20,
               dirspeed = wind, srtm_crop="",topograp=FALSE,cclRaster="")
allparks <- do.call("rbind",fit)

## SELECTION
## print the amount of Individuals selected.
## Check if the amount of Turbines is as requested.
selec6best <- selection(fit, Grid,2, TRUE, 6, "VAR")

## CROSSOVER
## u determines the amount of crossover points,
## crossPart determines the method used (Equal/Random),
## uplimit is the maximum allowed permutations
crossOut <- crossover(selec6best, 2, uplimit = 300, crossPart="RAN")

## MUTATION
## Variable Mutation Rate is activated if more than 2 individuals represent the
## current best solution.
mut <- mutation(a = crossOut, p = 0.3)

## TRIMTON
## After Crossover and Mutation, the amount of turbines in a windpark change
## and have to be corrected to the required amount of turbines.
mut1 <- trimton(mut = mut, nturb = 10, allparks = allparks,
               nGrids = AmountGrids, trimForce=FALSE)

## Get the new Grid-Ids and run a new fitness run.
getRectV <- get_grids(mut1, Grid)
fit <- fitness(selection = getRectV,referenceHeight = 100, RotorHeight=100,
               SurfaceRoughness=0.3,Polygon = Polygon1, resol1 = 200,rot=20,
               dirspeed = wind, srtm_crop="",topograp=FALSE,cclRaster="")
head(fit)

#> $`4,16,33,34,38,47,81,86,92,100`
#>        X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#> 4    700  100    97.76254      7710.578         100   1     20       4
#> 16  1100  300    97.76254      7710.578         100   1     20      16
#> 33   500  700    97.76254      7710.578           0   1     20      33
#> 34   700  700    97.76254      7710.578           0   1     20      34
#> 38  1500  700    97.76254      7710.578           0   1     20      38
#> 47  1300  900    97.76254      7710.578           0   1     20      47
#> 81   100 1700    97.76254      7710.578           0   1     20      81
#> 86  1100 1700    97.76254      7710.578           0   1     20      86
#> 92   300 1900    97.76254      7710.578           0   1     20      92
#> 100 1900 1900    97.76254      7710.578           0   1     20     100
#>     Parkfitness
#> 4      7710.578
#> 16     7710.578
#> 33     7710.578
#> 34     7710.578
#> 38     7710.578
#> 47     7710.578
#> 81     7710.578
#> 86     7710.578
#> 92     7710.578
#> 100    7710.578
#> 
#> $`14,16,32,50,72,75,79,81,88,94`
#>       X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#> 14  700  300    98.42905      7763.146         100   2     20      14
#> 16 1100  300    98.42905      7763.146           0   2     20      16
#> 32  300  700    98.42905      7763.146         100   2     20      32
#> 50 1900  900    98.42905      7763.146           0   2     20      50
#> 72  300 1500    98.42905      7763.146           0   2     20      72
#> 75  900 1500    98.42905      7763.146           0   2     20      75
#> 79 1700 1500    98.42905      7763.146           0   2     20      79
#> 81  100 1700    98.42905      7763.146           0   2     20      81
#> 88 1500 1700    98.42905      7763.146           0   2     20      88
#> 94  700 1900    98.42905      7763.146           0   2     20      94
#>    Parkfitness
#> 14    7763.146
#> 16    7763.146
#> 32    7763.146
#> 50    7763.146
#> 72    7763.146
#> 75    7763.146
#> 79    7763.146
#> 81    7763.146
#> 88    7763.146
#> 94    7763.146
#> 
#> $`7,21,24,31,32,44,58,70,86,90`
#>       X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#> 7  1300  100     89.2255      7037.257           0   3     20       7
#> 21  100  500     89.2255      7037.257         100   3     20      21
#> 24  700  500     89.2255      7037.257         100   3     20      24
#> 31  100  700     89.2255      7037.257           0   3     20      31
#> 32  300  700     89.2255      7037.257           0   3     20      32
#> 44  700  900     89.2255      7037.257           0   3     20      44
#> 58 1500 1100     89.2255      7037.257           0   3     20      58
#> 70 1900 1300     89.2255      7037.257         100   3     20      70
#> 86 1100 1700     89.2255      7037.257           0   3     20      86
#> 90 1900 1700     89.2255      7037.257           0   3     20      90
#>    Parkfitness
#> 7     7037.257
#> 21    7037.257
#> 24    7037.257
#> 31    7037.257
#> 32    7037.257
#> 44    7037.257
#> 58    7037.257
#> 70    7037.257
#> 86    7037.257
#> 90    7037.257
#> 
#> $`3,16,36,53,59,67,71,72,77,83`
#>       X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#> 3   500  100    89.37569      7049.103         200   4     20       3
#> 16 1100  300    89.37569      7049.103         100   4     20      16
#> 36 1100  700    89.37569      7049.103           0   4     20      36
#> 53  500 1100    89.37569      7049.103         100   4     20      53
#> 59 1700 1100    89.37569      7049.103           0   4     20      59
#> 67 1300 1300    89.37569      7049.103         100   4     20      67
#> 71  100 1500    89.37569      7049.103           0   4     20      71
#> 72  300 1500    89.37569      7049.103           0   4     20      72
#> 77 1300 1500    89.37569      7049.103           0   4     20      77
#> 83  500 1700    89.37569      7049.103           0   4     20      83
#>    Parkfitness
#> 3     7049.103
#> 16    7049.103
#> 36    7049.103
#> 53    7049.103
#> 59    7049.103
#> 67    7049.103
#> 71    7049.103
#> 72    7049.103
#> 77    7049.103
#> 83    7049.103
#> 
#> $`11,25,36,44,48,49,59,73,94,97`
#>       X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#> 11  100  300    94.02258      7415.605           0   5     20      11
#> 25  900  500    94.02258      7415.605           0   5     20      25
#> 36 1100  700    94.02258      7415.605           0   5     20      36
#> 44  700  900    94.02258      7415.605         100   5     20      44
#> 48 1500  900    94.02258      7415.605           0   5     20      48
#> 49 1700  900    94.02258      7415.605         100   5     20      49
#> 59 1700 1100    94.02258      7415.605           0   5     20      59
#> 73  500 1500    94.02258      7415.605           0   5     20      73
#> 94  700 1900    94.02258      7415.605           0   5     20      94
#> 97 1300 1900    94.02258      7415.605           0   5     20      97
#>    Parkfitness
#> 11    7415.605
#> 25    7415.605
#> 36    7415.605
#> 44    7415.605
#> 48    7415.605
#> 49    7415.605
#> 59    7415.605
#> 73    7415.605
#> 94    7415.605
#> 97    7415.605
#> 
#> $`2,8,31,36,62,73,84,85,90,100`
#>        X    Y EfficAllDir EnergyOverall AbschGesamt Run RotorR Rect_ID
#> 2    300  100     94.2324      7432.154         100   6     20       2
#> 8   1500  100     94.2324      7432.154           0   6     20       8
#> 31   100  700     94.2324      7432.154           0   6     20      31
#> 36  1100  700     94.2324      7432.154           0   6     20      36
#> 62   300 1300     94.2324      7432.154           0   6     20      62
#> 73   500 1500     94.2324      7432.154           0   6     20      73
#> 84   700 1700     94.2324      7432.154           0   6     20      84
#> 85   900 1700     94.2324      7432.154           0   6     20      85
#> 90  1900 1700     94.2324      7432.154         100   6     20      90
#> 100 1900 1900     94.2324      7432.154           0   6     20     100
#>     Parkfitness
#> 2      7432.154
#> 8      7432.154
#> 31     7432.154
#> 36     7432.154
#> 62     7432.154
#> 73     7432.154
#> 84     7432.154
#> 85     7432.154
#> 90     7432.154
#> 100    7432.154
#> 

# }