Polygon to Hexagonal Grid Tessellation

The function takes a Polygon and a sizing argument and creates a list with an indexed matrix with coordinates and a SpatialPolygons object, that consists of hexagonal grids

hexa_area(Polygon1, size, plotTrue = FALSE)

Arguments

Polygon1
The SpatialPolygons object

size
The side length of an hexagon

plotTrue
Should the object be plotted

Value

Returns a list with an indexed matrix of the point coordinates and a SpatialPolygons object of the hexagons

See also

Examples

library(spatstat)

#> Loading required package: spatstat.data

#> Loading required package: nlme

#> 
#> Attaching package: ‘nlme’

#> The following object is masked from ‘package:raster’:
#> 
#>     getData

#> Loading required package: rpart

#> 
#> spatstat 1.64-1       (nickname: ‘Help you I can, yes!’) 
#> For an introduction to spatstat, type ‘beginner’ 

#> 
#> Note: spatstat version 1.64-1 is out of date by more than 5 months; we recommend upgrading to the latest version.

#> 
#> Attaching package: ‘spatstat’

#> The following objects are masked from ‘package:raster’:
#> 
#>     area, rotate, shift

library(sp)
library(raster)
Polygon1 <- Polygon(rbind(c(4498482, 2668272), c(4498482, 2669343),
                          c(4499991, 2669343), c(4499991, 2668272)))
Polygon1 <- Polygons(list(Polygon1),1);
Polygon1 <- SpatialPolygons(list(Polygon1))
Projection <- "+init=epsg:3035"
proj4string(Polygon1) <- Projection
HexGrid <- hexa_area(Polygon1, 100, TRUE)

plot(HexGrid[[2]])