Plot an interpolated viewshed

Plot an interpolated view of the viewshed analysis

Usage

interpol_view(
  res,
  plot = TRUE,
  breakseq,
  breakform = NULL,
  plotDEM = FALSE,
  fun = mean,
  pal = NULL,
  ...
)

Arguments

res

The resulting list from viewshed

plot

Should the result be plotted? Default is TRUE

breakseq

The breaks for value plotting. By default, 5 equal intervals are generated.

breakform

If breakseq is missing, a sampling function to calculate the breaks, like quantile, fivenum, etc.

plotDEM

Plot the DEM? Default is FALSE

fun

Function used for rasterize. Default is mean

pal

A color palette

...

Arguments passed on to plot.

Value

An interpolated raster

See also

Examples

if (FALSE) {
library(raster)
library(sf)
Polygon1 <- sf::st_as_sf(sf::st_sfc(
  sf::st_polygon(list(cbind(
    c(4496482, 4496482, 4499991, 4499991, 4496482),
    c(2666272, 2669343, 2669343, 2666272, 2666272)))),
  crs = 3035
))
DEM_meter <- getDEM(Polygon1)

turbloc = st_sample(DEM_meter[[2]], 10, type = "random");
res <- viewshed(r = DEM_meter[[1]], shape=DEM_meter[[2]], 
                turbine_locs = turbloc,  h1=1.8, h2=50)
interpol_view(res, plotDEM = TRUE)

interpol_view(res, breakseq = seq(0,max(colSums(res$Result)),1))
interpol_view(res, plotDEM = FALSE, breakform = quantile)
interpol_view(res, breakform = factor)

## Different color palettes
interpol_view(res, plotDEM = TRUE, pal=topo.colors)
interpol_view(res, plotDEM = TRUE, pal=colorRampPalette(c("white","purple")))

## ... Arguments are past on to the raster plot method
interpol_view(res, plotDEM = TRUE, alpha=0.5)
interpol_view(res, plotDEM = FALSE, breakseq = seq(0,10,1), colNA="black")

}