Enumerate the Combinations or Permutations of the Elements of a Vector

permutations enumerates the possible permutations. The function is forked and minified from gtools::permutations

permutations(n, r, v = 1:n)

Arguments

n
Size of the source vector

r
Size of the target vectors

v
Source vector. Defaults to 1:n

Value

Returns a matrix where each row contains a vector of length r.

References

Venables, Bill. "Programmers Note", R-News, Vol 1/1, Jan. 2001. https://cran.r-project.org/doc/Rnews/

See also

Author

Original versions by Bill Venables Bill.Venables@cmis.csiro.au. Extended to handle repeats.allowed by Gregory R. Warnes greg@warnes.net.