Randomize the location of a single turbine

Perform a random search for a single turbine, to further optimize the output of the wind farm layout.

Usage

random_search_single(result, Polygon1, n = 20, Plot = FALSE, max_dist = 2.2)

Arguments

result

The resulting matrix of the function genetic_algorithm

Polygon1

The considered area as SpatialPolygon, SimpleFeature Polygon or coordinates as matrix/data.frame

n

The number of random searches to be performed. Default is 20.

Plot

Should the random search be plotted? Default is FALSE

max_dist

A numeric value multiplied by the rotor radius to perform collision checks. Default is 2.2

Value

Returns a list

See also